% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{apply_binary_mask}
\alias{apply_binary_mask}
\title{Create a binary group value from a set of logical columns.}
\usage{
apply_binary_mask(data, mask, column)
}
\arguments{
\item{data}{data.frame}

\item{mask}{binary mask for each column (see \code{\link{create_binary_mask}})}

\item{column}{name of the column that will contain the group value}
}
\description{
Intersection of columns will be encoded as a binary value (1 column => 1 bit)
}
\seealso{
Other upset: 
\code{\link{create_binary_groups}()},
\code{\link{create_binary_mask}()},
\code{\link{get_labels_from_binary}()},
\code{\link{upset_plot}()}
}
\concept{upset}
