% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{as_mapping}
\alias{as_mapping}
\title{Create a mapping to recode variable value to labels}
\usage{
as_mapping(codes, labels)
}
\arguments{
\item{codes}{vector of values as stored in the database}

\item{labels}{vector of labels}
}
\description{
Create a mapping to recode variable value to labels
}
