% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{axis.week}
\alias{axis.week}
\title{Axis for yearweek data}
\usage{
axis.week(side, ww, mode = c("ticks", "year", "week", "weeks", "\%\%"),
  format = "yw", col.yw = "yw", sep = "s", century = T, ticks = NULL,
  ...)
}
\arguments{
\item{side}{side of the axeis}

\item{ww}{data.frame(), with cols wid=week index, [col.yw]=yearweek value (@see makeWeekIndex)}

\item{mode}{"ticks" follow ticks, "year" each 1st week of each year, "week" (or weeks) some given weeks, "\%\%" root of modulo}

\item{format}{"yw" pretty week format, "w"=only week number, NULL=disable}

\item{col.yw}{name of the yearweek value colum in ww}

\item{sep}{separator between year and week number}

\item{century}{use century for year number}

\item{ticks}{if mode=weeks vector of week (1-53) number, if mode="\%\%" modulo to use}
}
\description{
Create an axis from weekly data (when the plot use week index, and you want to label axe with classical week nubmers)
}
