% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset.R
\name{base_onset_design}
\alias{base_onset_design}
\title{Base onset design, used for incidence computation}
\usage{
base_onset_design()
}
\value{
quosude expression used to compute onssed
}
\description{
See \code{\link{compute_onset}}
}
