% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{box.sd}
\alias{box.sd}
\title{Make a rectangle usefull for boxplot}
\usage{
box.sd(x, up, low, k = 0.25, col = "grey", border = NA)
}
\arguments{
\item{x}{x value}

\item{up}{upper value}

\item{low}{lower value}

\item{k}{rectangle width around the x}
}
\description{
Make a rectangle usefull for boxplot
}
