% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{calc.age}
\alias{calc.age}
\title{Age of participants from a Year-month birth date}
\usage{
calc.age(ym, time)
}
\arguments{
\item{ym}{year-month vector}

\item{time}{reference time}
}
\description{
Age of participants from a Year-month birth date
}
