% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_design.R
\name{calc.conf.int}
\alias{calc.conf.int}
\title{Compute confidence interval for rate
using poisson distribution law}
\usage{
calc.conf.int(inc, v, type = "crude", unit = 1)
}
\arguments{
\item{inc}{data.frame from incidence estimator}

\item{v}{column name containing syndrom count}

\item{type}{crude to create rate column name}

\item{unit}{if rate need to be rescale}
}
\description{
Compute confidence interval for rate
using poisson distribution law
}
