% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{calc.season.fixed}
\alias{calc.season.fixed}
\title{Compute a season index of each year week}
\usage{
calc.season.fixed(inc, col.yw = "yw", col.stamp = NULL,
  date.start = "09-01", calc.index = TRUE)
}
\arguments{
\item{inc}{data.frame() data}

\item{col.yw}{name of the column containing the yearweek number}

\item{col.stamp}{name of the column containing the week stamp of each yearweek (@see WeekStamp), if not provided compute it}

\item{date.start}{month-day date to take as the start of each season}

\item{calc.index}{if TRUE add an index column}
}
\value{
inc with 'season.year' (year of the start ) and 'season.index' (index of the week in the season) columns
}
\description{
For each week compute the number of the week inside the "season", using an arbitrary start date for the season
Based on a fixed date in the year (to take account of the week lag for some years)
Conventionnaly we took the first of september of each year as the season's start
}
