% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{calc.season}
\alias{calc.season}
\title{Return the season number of a yearweek or a Date
Currently the season is defined from september to august of the next year and the season number is the year of the september month
(i.e. the first year of the season)}
\usage{
calc.season(d)
}
\description{
Return the season number of a yearweek or a Date
Currently the season is defined from september to august of the next year and the season number is the year of the september month
(i.e. the first year of the season)
}
