% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_design.R
\name{calc_adjusted_incidence}
\alias{calc_adjusted_incidence}
\title{Compute crude & adjusted incidence from a count data.frame
This function is shared by all incidence computation methods, once active participants & syndrome counts are provided}
\usage{
calc_adjusted_incidence(count.week, design, syndromes, output)
}
\arguments{
\item{count.week}{data.frame(), count for a week. Count of active participants is in "active" column}

\item{design}{design structure (defining strata and reference data like population in each strata)}

\item{syndromes}{char() list of column names for each syndrome count}

\item{output}{char() list of requested output "inc"=global incidence, "zlow"=lower geo level incidence, "age" by age incidence,}
}
\description{
Ouput columns are ([syndrome] is the column with logical value for presence of one syndrome in the weekly data, for example "ili"):
 - active : count of active participants
 - [syndrome] : count for this syndrome (or variable)
 - [syndrome].crude : crude incidence (without adjustment, count/active participant ratio)
 - [syndrome].adj : Adjusted incidence (taking into account of stratification design)
 - [syndrome].crude.(upper|lower) : upper & lower bound of CI95 for crude incidence (Exact poisson CI)
 - [syndrome].adj.(upper|lower) : upper & lower boud of CI95 for ajusted incidence (DKES method)
}
