% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{calc_season}
\alias{calc_season}
\title{Return the season number of a yearweek or a Date}
\usage{
calc_season(d)
}
\arguments{
\item{d}{date to compute season number from}
}
\description{
Currently the season is defined from september to august of the next year and the season number is the year of the september month
(i.e. the first year of the season).
}
