% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{calc_season_of_date}
\alias{calc_season_of_date}
\title{Compute season number of a date form historical seasons definition}
\usage{
calc_season_of_date(date)
}
\description{
Compute season number of a date form historical seasons definition
}
