% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_utils.R
\name{calc_weekly_delay}
\alias{calc_weekly_delay}
\title{Compute delay between 2 surveys for each participant}
\usage{
calc_weekly_delay(weekly, time.col)
}
\arguments{
\item{weekly}{data.frame with at least (person_id, [time.col]) columns}

\item{time.col}{column containg time value to compute delay from}
}
\description{
weekly should be ordered by person_id then timestamp !
}
