% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_utils.R
\name{calc_weekly_order}
\alias{calc_weekly_order}
\title{Compute order of weekly response}
\usage{
calc_weekly_order(weekly)
}
\arguments{
\item{weekly}{weekly data.frame()}
}
\description{
Compute order of weekly response
}
