% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{can_use_country}
\alias{can_use_country}
\title{Check if the platform is allowed to use "country" value}
\usage{
can_use_country(country)
}
\value{
TRUE if country is defined and platform can use country
}
\description{
raise an error if the platform is not configured to use country value
}
