% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{check_list_mapping}
\alias{check_list_mapping}
\title{Check if two named list are compatible to merge}
\usage{
check_list_mapping(new, old, raise, only.errors = TRUE)
}
\arguments{
\item{new}{list()}

\item{old}{list()}

\item{raise}{function(type, value, problem, message)}

\item{only.errors}{only raise errors}
}
\description{
To be compatible with template, mapping should follow these rules:
\itemize{
  \item{only one value is mapped to a name}
  \item{new can add new name with new value, not already in old}
  \item{new cannot redefine a value with a new name (unless allow_override, not currently supported)}
}
}
\details{
Overriding is allowed wraping the value with \code{\link{override}} function
}
\examples{
\dontrun{
new = list("var1"="Q1", "var2"="Q2", "var3"="Q3", "var4"="Q1", "var5"="Q4")
old = list("var0"="Q0", "var1"="Q1", "var2"="Q4")
check_list_mapping(new, old, raise=raise)
# Returs list of errors

new = list("var1"="Q1", "var3"=override("Q2"))
old = list("var2"="Q2")
check_list_mapping(new, old, raise=raise) # Only throw warning
}
# The checks are transmitted to a provided function \\code{\\link{raise()}}
}
