% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{check_season_dates}
\alias{check_season_dates}
\title{Check if provided date bounds are compatible with a season}
\usage{
check_season_dates(def, date)
}
\arguments{
\item{def}{season definition returned by \code{\link{season_definition}()}}

\item{date}{list() list of min, max requested dates}
}
\description{
Check if provided date bounds are compatible with a season
}
