% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{check_survey_template}
\alias{check_survey_template}
\title{Check if a survey definition is compatible with a survey template}
\usage{
check_survey_template(template, mapping, recodes, only.errors = TRUE)
}
\arguments{
\item{template}{template name}

\item{mapping}{new question mapping to test}

\item{recodes}{new recoding to test}

\item{only.errors}{boolean if TRUE only report errors}
}
\value{
list()
}
\description{
Check if a survey definition is compatible with a survey template
}
