% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{check_unique}
\alias{check_unique}
\title{Check if list values are unique}
\usage{
check_unique(x)
}
\arguments{
\item{x}{vector or list of values to check for unicity}
}
\description{
Check if list values are unique
}
\keyword{internal}
