% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_loader.R
\name{complete_intake_strategy}
\alias{complete_intake_strategy}
\alias{complete_intake}
\title{Complete intake according to platform's strategy}
\usage{
complete_intake_strategy(data, intake, ...)

complete_intake(data, intake, intake.columns, geo = NULL,
  max.year = NA)
}
\arguments{
\item{data}{data.frame() survey data with all participants should have at least (timestamp, person_id) columns. For example weekly survey data}

\item{intake}{data.frame() intake survey data loaded on the same period as data}

\item{intake.columns}{columns to load with intake}

\item{geo}{geo levels to load with intake}
}
\value{
intake with extra intake loaded from previous season
}
\description{
Complete intake according to platform's strategy

Complete intake for a survey using intake from the past seasons
}
