% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_loader.R
\name{complete_intake_strategy}
\alias{complete_intake_strategy}
\title{Complete intake according to platform's strategy}
\usage{
complete_intake_strategy(data, intake, ...)
}
\arguments{
\item{data}{data.frame() survey data with all participants (typically weekly), see \code{\link{complete_intake}}}

\item{intake}{data.frame() intake data loaded on same period as data}

\item{...}{other parameters to pass to \code{\link{complete_intake}}}
}
\description{
Complete intake according to platform's strategy
}
