% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset.R
\name{compute_onset}
\alias{compute_onset}
\title{Compute the onset date from a design object}
\usage{
compute_onset(weekly, design)
}
\arguments{
\item{weekly}{weekly data}

\item{design}{design object,}
}
\value{
weekly
}
\description{
Evaluate the `design` expression to compute the onset column.
The expression is evaluated using weekly data and quosure environment;
}
\details{
if design is FALSE, does nothing, return the weekly. This can be used if the `onset` column is already available in
the provided weekly data.frame.
}
