% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_loader.R
\name{compute_weekly_syndromes}
\alias{compute_weekly_syndromes}
\title{Create syndrome columns in the weekly data}
\usage{
compute_weekly_syndromes(
  intake,
  weekly,
  health.status = TRUE,
  keep.status = FALSE,
  regroup.syndromes = TRUE,
  provider = NULL
)
}
\arguments{
\item{intake}{intake data.frame()}

\item{weekly}{weekly data.frame()}

\item{health.status}{bool use the default health status computed using InfluenzaNet default strategy (used in the website)}

\item{keep.status}{bool keep the original health status (from InfluenzaNet view), renamed to "status.old"}

\item{regroup.syndromes}{bool use syndrome grouping (recode syndrome list for Influenzanet's health status list) to a simplier list}

\item{provider}{function(weekly,intake) returning a data.frame to be merged into weekly (using "id" weekly's column as merge key), useable to compute custom syndromes}
}
\description{
This function use syndrome classifier (called provider)
}
