% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{create_binary_groups}
\alias{create_binary_groups}
\title{Create new groups from a set of columns}
\usage{
create_binary_groups(data, groups, trans = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{groups}{list of group, each group is a vector of names of columns belonging to this group}

\item{trans}{translate label if TRUE see \code{\link{get_labels_from_binary}}}
}
\description{
Names of new groups will be the name of groups entry
}
\seealso{
Other upset: 
\code{\link{apply_binary_mask}()},
\code{\link{create_binary_mask}()},
\code{\link{get_labels_from_binary}()},
\code{\link{upset_plot}()}
}
\concept{upset}
