% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{create_participant_weight}
\alias{create_participant_weight}
\title{Define participants weight from general population to compute adjusted frequency}
\usage{
create_participant_weight(
  age.categories,
  year.pop,
  groups,
  strata,
  weekly,
  intake
)
}
\arguments{
\item{age.categories}{age breaks vector (strictly increasing and unique values)}

\item{year.pop}{year of the population to load}

\item{groups}{list of weekly column to use as weight group}

\item{strata}{list of intake columns to use as stratification to compute weight.. Must be one of pop (age.cat or gender)}

\item{weekly}{weekly data}

\item{intake}{intake data}
}
\value{
data.frame() with columns (person_id, [groups...], [strata...], weight)
}
\description{
Weights can be stratified and grouped : weights are computed by strata over each group
For example : strata=[age, gender] and group = week, will allow to compute weekly frequency adjusted by age and gender
}
\details{
Strata can be either 'age.cat' or 'gender', since provided population are by age and gender. To have more strata will need to have
general population data for each strata.
}
