% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{create_path}
\alias{create_path}
\title{Internal function create the path from the current paths prefix & suffix}
\usage{
create_path()
}
\arguments{
\item{p}{path to add}
}
\description{
Internal function create the path from the current paths prefix & suffix
}
