% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{create_survey_definition}
\alias{create_survey_definition}
\title{Create a survey definition}
\usage{
create_survey_definition(
  mapping,
  labels = NULL,
  codes = NULL,
  recodes = list(),
  template = NULL,
  only.errors = TRUE
)
}
\arguments{
\item{mapping}{list() variable (name) to DB column (value) mapping}

\item{labels}{list() named list of labels}

\item{codes}{list() of codes **deprecated**}

\item{recodes}{list() list of recoding}

\item{template}{character}

\item{only.errors}{if TRUE only check for errors}
}
\value{
list(mapping, labels, recodes, checks)
}
\description{
Create and check survey structure and import from template if provided.
This function is mainly used internally by \code{\link{platform_define_survey}}. Unless you want to test definition you should not need this function.
}
\seealso{
\code{\link{platform_define_survey}}
}
