% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_utils.R
\name{create_syndrom_columns}
\alias{create_syndrom_columns}
\title{Create syndrom columns from factor encoded column}
\usage{
create_syndrom_columns(weekly, column)
}
\arguments{
\item{weekly}{weekly data.frame()}

\item{column}{column containing syndrom name}
}
\description{
Create one boolean column for each level of the given column
}
