% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.r
\name{cut_age}
\alias{cut_age}
\title{Cut age values into categories using breaks and defining pretty levels labels}
\usage{
cut_age(age, age.categories)
}
\arguments{
\item{age}{vector of age values}

\item{age.categories}{list of breaks (@seealso cut)}
}
\description{
Cut age values into categories using breaks and defining pretty levels labels
}
