% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{db_equal}
\alias{db_equal}
\title{Create an equal criteria}
\usage{
db_equal(field, value, quote = NA, several = FALSE)
}
\arguments{
\item{field}{name of field (unquoted)}

\item{value}{value to compare}

\item{quote}{if TRUE quote the value, NA = detect}

\item{several}{if TRUE accept to compare several values}
}
\description{
Create an equal criteria
}
