% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset.R
\name{default_onset_design}
\alias{default_onset_design}
\title{Create default expression to compute onset design for episodes}
\usage{
default_onset_design(delay_symptome_max = 15)
}
\arguments{
\item{delay_symptome_max}{integer max delay to use `fever.start` or `sympt.start` instead of survey date}
}
\description{
Create default expression to compute onset design for episodes
}
