% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{default_season_dates}
\alias{default_season_dates}
\title{Generate Default dates bound of a season using 1st of September
Season are defined from 1s September of season year, to next 31 august}
\usage{
default_season_dates(season, start = "09-01")
}
\arguments{
\item{season}{season number}

\item{start}{cutting date as month-day, default is 1st of September}
}
\description{
Generate Default dates bound of a season using 1st of September
Season are defined from 1s September of season year, to next 31 august
}
