% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_design.R
\name{design_incidence}
\alias{design_incidence}
\title{Describe the stratification to use for incidence computation (parameter design)}
\usage{
design_incidence(
  age.categories,
  year.pop,
  geo,
  geo_column = NULL,
  geo_area = NULL,
  use.gender = FALSE,
  ...
)
}
\arguments{
\item{age.categories}{list of breaks}

\item{year.pop}{year of the population to fetch}

\item{geo}{geographic level id (see geography & geo.levels in platform config.). could be character or structure return by geo_level function}

\item{geo_column}{name of the geographic column  (value return by strata.call), geo_level's column name by default}

\item{geo_area}{id of geographic area to keep at the geographic level}

\item{use.gender}{use gender as strata}

\item{...}{other parameters to include in the design}
}
\value{
design.incidence object
}
\description{
Create a structure usable by incidence computation functions.
The returned structure embed all data & metadata about stratification
Only handle strata on age and geo level for instance
By convention age group column will be names "age.cat"
}
