% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{design_weight}
\alias{design_weight}
\title{Create a design for adjustment computation}
\usage{
design_weight(data, col_weight = "weight")
}
\arguments{
\item{data}{dataset to adjust on}

\item{col_weight}{name of the column containing weight}
}
\value{
survey::svydesign object
}
\description{
Create a design for adjustment computation
}
