% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episode_build}
\alias{episode_build}
\title{Build Episodes data from survey data}
\usage{
episode_build(design, intake, weekly, syndrome.column)
}
\arguments{
\item{design}{a design structure holding computation parameters, \code{\link{episode_design}}}

\item{intake}{intake data}

\item{weekly}{weekly data}

\item{syndrome.column}{name of the boolean column indicating if each weekly fit the syndrome definition or not (TRUE is yes)}
}
\value{
environment containing intake, weekly, participants
}
\description{
Build Episodes data from survey data
}
