% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episode_compute}
\alias{episode_compute}
\title{Compute Episodes}
\usage{
episode_compute(.env, syndrome.column, design, .progress = TRUE,
  local = FALSE)
}
\arguments{
\item{.env}{environment result of episode_prepare_data()}

\item{syndrome.column}{name of the column containing episode data}

\item{design}{study params (episode_design)}

\item{.progress}{optional progress bar structure}

\item{local}{if TRUE do not modify .env, only returns episodes}
}
\value{
data.frame (episodes)
}
\description{
Compute Episodes
}
