% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episode_compute_ariza}
\alias{episode_compute_ariza}
\title{Compute episodes using M Ariza's algorithm}
\usage{
episode_compute_ariza(weekly, syndrome.column, params, .progress = NULL)
}
\arguments{
\item{weekly}{data.frame of weekly data}

\item{syndrome.column}{name of the syndrome column to use}

\item{params}{episode_design}

\item{.progress}{progress_estimated style progress bar, nothing if NULL
Find episodes, identify episodes sequences for one person_id
Returns a data.frame() with weekly row id (column "id") & episode number}
}
\description{
Compute episodes using M Ariza's algorithm
}
