% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episode_design}
\alias{episode_design}
\title{Create a design definition for episodes computation}
\usage{
episode_design(delay_episode = 15, participants = list(),
  method = "ariza", onset = NULL, strategies = NULL)
}
\arguments{
\item{delay_episode}{int maximum delay to consider two sequential responses are in the same episode}

\item{participants}{list of rules to apply for participants selection will be passed to \code{\link{episode_select_participants}} as `rules` arguments}

\item{onset}{onset design expression or false, design parameter passed to \code{\link{compute_onset}}}

\item{strategies}{not implemented yet}
}
\value{
episodes design data structure
}
\description{
A "design" is a set of information about a process, it's a structure embedding parameters to compute the episodes
}
