% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_fusion.R
\name{episode_fusion.simple_strategy}
\alias{episode_fusion.simple_strategy}
\title{Merge episodes using a simple function as merge strategy}
\usage{
\method{episode_fusion}{simple_strategy}(strategy, weekly, episode.column, ...)
}
\arguments{
\item{strategy}{strategy definition structure, created by \code{\link{episode_strategy}}}

\item{weekly}{weekly data.frame for one episode}

\item{episode.column}{character string name of column containing episode}

\item{...}{other arguments (not used)}
}
\value{
data.frame() one row by person_id, episode, with selected value for each variable registred in strategy
}
\description{
Merge episodes using a simple function as merge strategy
}
\seealso{
\code{\link{episode_fusion}}
}
