% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_fusion.R
\name{episode_fusion_strategy}
\alias{episode_fusion_strategy}
\title{Compute episode dataset}
\usage{
episode_fusion_strategy(weekly, design, episode.column = "episode")
}
\arguments{
\item{weekly}{weekly data with episode column (typically computed by \code{\link{episode_compute}})}

\item{design}{study params, structure returned by \code{\link{episode_design}}}

\item{episode.column}{name of the column containing episode identifier}
}
\description{
This function will keep only one weekly by disease "episode" previously computed in the column in \code{episode.column} parameter
Variables of all weekly in the same episode will
}
\seealso{
Other episodes: 
\code{\link{episode_build}()},
\code{\link{episode_compute}()},
\code{\link{episode_design}()},
\code{\link{episode_fusion.simple_strategy}()},
\code{\link{episode_fusion.worst_strategy}()},
\code{\link{episode_fusion}()},
\code{\link{episode_prepare_data}()},
\code{\link{episode_select_participants}()},
\code{\link{episode_strategy}()}
}
\concept{episodes}
