% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episode_prepare_data}
\alias{episode_prepare_data}
\title{Prepare data and select participants using given rules}
\usage{
episode_prepare_data(design, intake, weekly)
}
\arguments{
\item{design}{episode design data structure (result of \code{episode_design()})}

\item{intake}{data.frame intake data}

\item{weekly}{data.frame weekly data}
}
\value{
environment environment containing selected data (weekly, intake, participants, selections)
}
\description{
Prepare data and select participants using given rules
}
\details{
The environment contains several data.frames()
- participants : one row for each participants and rule results (TRUE/FALSE) for each participants, column "keep" indicate final state of participant
- selections : count of selected participants at each step (after applying each rule)
- intake : the intakes for selected participants
- weekly ! the weekly for selected participants
}
