% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episode_select_participants}
\alias{episode_select_participants}
\title{Select participants based on a list of rules}
\usage{
episode_select_participants(weekly, intake, rules)
}
\arguments{
\item{weekly}{weekly data.frame weekly data}

\item{intake}{intake data.frame intake data}

\item{rules}{list of rule definition (see details)}
}
\value{
data.frame() participants state for each rules, selections count at each step in attributes
}
\description{
Select participants based on a list of rules
}
\details{
Each rule is a element of the list using 3 forms
\itemize{
 \item{One character value, the name of the rule (without parameters)}
 \item{a named list element, rule name as name, parameters as values (single value or list of named parameters)}
 \item{a anonymous element, rule name is given in the `name` entry}
}

Known rules:
\describe{
\item{has.intake}{Only participants with at least one intake}
\item{has.weekly}{Only participants with at least one weekly}
\item{min.survey}{Participants with at least [value] survey during the season, one parameter as single int value}
\item{has_before}{Participants with at least one survey before the given date, one parameter a date}
\item{has_after}{Participants with at least one survey before the given date, one parameter a date}
\item{has_between}{Participants with at least one survey between the given dates, "start" and "end" param are expected}
\item{remove_domtom}{Remove french overseas - french specific }
}
}
\examples{

rules = list(
"has.intake",
"has.weekly",
"min.survey"=3,
"has_before"="2019-01-01",
"has_between"=list(start="2019-01-02",end="2019-03-01")
)

}
\seealso{
Other episodes: 
\code{\link{episode_build}()},
\code{\link{episode_compute}()},
\code{\link{episode_design}()},
\code{\link{episode_fusion.simple_strategy}()},
\code{\link{episode_fusion.worst_strategy}()},
\code{\link{episode_fusion_strategy}()},
\code{\link{episode_fusion}()},
\code{\link{episode_prepare_data}()},
\code{\link{episode_strategy}()}
}
\concept{episodes}
