% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_fusion.R
\name{episode_strategy}
\alias{episode_strategy}
\title{Define a strategy to merge data of the same episode of a participant}
\usage{
episode_strategy(type, ...)
}
\arguments{
\item{type}{name of the strategy to apply}

\item{...}{variable names in weekly (columns) & parameters to apply (see details)}
}
\description{
Define a strategy to merge data of the same episode of a participant
}
\details{
Parameters:

The parameters will depends on the type of strategy

For a simple strategy (with no parameter attached to each variable), it will be the name of column to apply the strategy to

For a complex strategy it will be the variable name as argument name passed to this function and the parameters for the variables as the argument value

Type of strategies:

Several kind of strategies are useasble:
\describe{
 \item{min}{Minimal value of the variable for the episode}
 \item{max}{Maximal value of the variable for the episode}
 \item{any}{Any value TRUE will output TRUE}
 \item{mean}{Average value of variable for the episode}
 \item{first}{Take the first available value (not NA) in each episode surveys of participant}
 \item{last}{Take the last available value (not NA) in each episode surveys of participant}
 \item{worst}{Worst Strategy algorithm, see details}
}

General warning:

You should be very careful by defining strategy on what data it will be applied. Some strategies will depends on how variables has been
recoded or not. Some checks are done but they are not perfects.

Worst strategy:

Worst strategy will expect
}
\examples{
\dontrun{
# Create an "any" strategy for some columns
episode_strategy("any", "chills", "cough", "fever")

# Define a worst strategy for "take.temp" column in weekly, and give the response order to use

episode_strategy("worst", "take.temp"=c('Yes','DontKnow','No'))
}
}
