% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\docType{data}
\name{FEVER.CODES}
\alias{FEVER.CODES}
\title{List of coded values for fever
names are pretty labels, values coded}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
FEVER.CODES
}
\description{
List of coded values for fever
names are pretty labels, values coded
}
\keyword{datasets}
