% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{format_week}
\alias{format_week}
\title{Format a yearweek value to a human readable format}
\usage{
format_week(w, sep = "s", century = T)
}
\arguments{
\item{w}{vector of yearweek values}

\item{sep}{separator to use}

\item{century}{use century for year value}
}
\description{
Format a yearweek value to a human readable format
}
