% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{freq_bool_by}
\alias{freq_bool_by}
\title{Compute frequency of a set of logical variables by group
Adjusted frequency can be computed by providing a design object with computed weights}
\usage{
freq_bool_by(data, vars, by, design = NULL)
}
\arguments{
\item{data}{data.frame containing data}

\item{vars}{list of variable names to compute frequency on. They must be logical}

\item{by}{grouping variable}

\item{design}{`svydesign` object, if provided will compute adjusted frequency using the design (data should contains the weight column as described in design)}
}
\description{
Compute frequency of a set of logical variables by group
Adjusted frequency can be computed by providing a design object with computed weights
}
