% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{freq_var_by}
\alias{freq_var_by}
\title{Compute frequency by group of a qualitative variable of a data.frame()}
\usage{
freq_var_by(data, variable, by = "yw", complete = TRUE)
}
\arguments{
\item{data}{data.frame()}

\item{variable}{variable to compute frequency on (for each level)}

\item{by}{grouping variables (final output)}

\item{complete}{if TRUE missing levels in groups will be completed with zero frequency}
}
\value{
data.frame()
  Returned columns:
  \describe{
    \item{level}{level of the variable}
    \item{n}{frequency of the level in the group}
    \item{n_total}{Total of non missing value in for the variable in the group}
    \item{prop_raw}{Proportion of the level}
    \item{group...}{One column of each group, with the value of the current group}
  }
}
\description{
Compute frequency of each level (or value) of the variable, by another grouping variable(s)
}
