% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{freq_vars_by}
\alias{freq_vars_by}
\title{Frequency of several variables by group}
\usage{
freq_vars_by(data, vars, by)
}
\arguments{
\item{data}{data.frame() data}

\item{vars}{list of variables to compute frequency on}

\item{by}{name of grouping variable}
}
\value{
data.frame()
  Returned columns:
  \describe{
    \item{variable}{variable name},
    \item{level}{level of the variable}
    \item{n}{frequency of the level in the group}
    \item{n_total}{Total of non missing value in for the variable in the group}
    \item{prop_raw}{Proportion of the level}
    \item{group...}{One column of each group, with the value of the current group}
  }
  An attribute "missing" contains missing count for each variable by group
}
\description{
Frequency of several variables by group
}
