% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_base_level}
\alias{geo_base_level}
\title{Get base level of geographic tables (lowest level)}
\usage{
geo_base_level()
}
\description{
Get base level of geographic tables (lowest level)
}
\seealso{
Other geography: 
\code{\link{geo_base_table}()},
\code{\link{geo_definition}()},
\code{\link{geo_hierarchy}()},
\code{\link{geo_is_type}()},
\code{\link{geo_level_nav}()},
\code{\link{geo_level_table}()},
\code{\link{geo_level}()},
\code{\link{get_geo_join}()},
\code{\link{load_geo_zone}()},
\code{\link{load_population_age}()},
\code{\link{load_population}()}
}
\concept{geography}
