% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_base_table}
\alias{geo_base_table}
\title{Get Base geographic tables
It should contains all levels, describe all upper levels from the base level}
\usage{
geo_base_table()
}
\description{
Get Base geographic tables
It should contains all levels, describe all upper levels from the base level
}
\seealso{
Other geography: 
\code{\link{geo_base_level}()},
\code{\link{geo_definition}()},
\code{\link{geo_hierarchy}()},
\code{\link{geo_is_type}()},
\code{\link{geo_level}()},
\code{\link{geo_level_nav}()},
\code{\link{geo_level_table}()},
\code{\link{get_geo_join}()},
\code{\link{load_geo_zone}()},
\code{\link{load_population}()},
\code{\link{load_population_age}()}
}
\concept{geography}
