% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_check_level}
\alias{geo_check_level}
\title{Check if level exists}
\usage{
geo_check_level(level, error = TRUE)
}
\arguments{
\item{level}{character vector of level names}

\item{error}{if TRUE raise an error if level doesnt exists}
}
\description{
Check if level exists
}
