% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_column}
\alias{geo_column}
\title{Get geographic column name for a given level}
\usage{
geo_column(geo)
}
\arguments{
\item{geo}{geographic level name}
}
\description{
Get geographic column name for a given level
}
