% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_definition}
\alias{geo_definition}
\title{Get geographic structure description}
\usage{
geo_definition(name = NULL)
}
\arguments{
\item{name}{geographic level name}
}
\description{
Get geographic structure description
}
\seealso{
Other geography: 
\code{\link{geo_base_level}()},
\code{\link{geo_base_table}()},
\code{\link{geo_hierarchy}()},
\code{\link{geo_is_type}()},
\code{\link{geo_level}()},
\code{\link{geo_level_nav}()},
\code{\link{geo_level_table}()},
\code{\link{get_geo_join}()},
\code{\link{load_geo_zone}()},
\code{\link{load_population}()},
\code{\link{load_population_age}()}
}
\concept{geography}
