% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_hierarchy}
\alias{geo_hierarchy}
\title{Get a hierarchy path of geographic levels}
\usage{
geo_hierarchy(hierarchy = NULL)
}
\arguments{
\item{hierarchy}{name of the hierarchy path to use}
}
\description{
A hierarchy path is a list of ordered geographic levels from the lowest to the uppest
}
\details{
Several hierarchy could be defined to handle several kind of levels aggregation for example

Depending of the context several kind of geographic levels can be used
Inside a hierarchy, all levels should be totaly nested (all sublevel should be mapped to ONE upper level)
}
\seealso{
[geo_levels()]

Other geography: 
\code{\link{geo_base_level}()},
\code{\link{geo_base_table}()},
\code{\link{geo_definition}()},
\code{\link{geo_is_type}()},
\code{\link{geo_level_nav}()},
\code{\link{geo_level_table}()},
\code{\link{geo_level}()},
\code{\link{get_geo_join}()},
\code{\link{load_geo_zone}()},
\code{\link{load_population_age}()},
\code{\link{load_population}()}
}
\concept{geography}
