% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_is_type}
\alias{geo_is_type}
\title{Allow to define "type" of geographic levels & area}
\usage{
geo_is_type(geo, type, code)
}
\arguments{
\item{geo}{geographic levelv}

\item{type}{feature to test for level's codes}

\item{code}{list of code of the level's area to test for the feature given in type}
}
\value{
logical vector of length of code
}
\description{
For example, it can be used to differenciate some code, corresponding to overseas area to mainland codes
The use of the function is not standardized yet.
}
\details{
TODO: should be a more generic function allowing to query "feature" about
Not exported currently, only to be used by platform defined functions by default returns TRUE for all queried area
}
\seealso{
Other geography: 
\code{\link{geo_base_level}()},
\code{\link{geo_base_table}()},
\code{\link{geo_definition}()},
\code{\link{geo_hierarchy}()},
\code{\link{geo_level}()},
\code{\link{geo_level_nav}()},
\code{\link{geo_level_table}()},
\code{\link{get_geo_join}()},
\code{\link{load_geo_zone}()},
\code{\link{load_population}()},
\code{\link{load_population_age}()}
}
\concept{geography}
