% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_level}
\alias{geo_level}
\title{Create a geographic level object}
\usage{
geo_level(level, hierarchy = NULL, ...)
}
\arguments{
\item{level}{geographic level name (for example "nuts1", "nuts2")}

\item{hierarchy}{name of a level hierarchy}

\item{...}{extra information to store with level information}
}
\description{
The available levels depends on the platform configuration
}
\seealso{
Other geography: 
\code{\link{geo_base_level}()},
\code{\link{geo_base_table}()},
\code{\link{geo_definition}()},
\code{\link{geo_hierarchy}()},
\code{\link{geo_is_type}()},
\code{\link{geo_level_nav}()},
\code{\link{geo_level_table}()},
\code{\link{get_geo_join}()},
\code{\link{load_geo_zone}()},
\code{\link{load_population}()},
\code{\link{load_population_age}()}
}
\concept{geography}
