% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_level_nav}
\alias{geo_level_nav}
\title{Navigate through geographic levels (get the upper level, the lower level, ...)}
\usage{
geo_level_nav(geo, side, hierarchy = NULL)
}
\arguments{
\item{geo}{the current level}

\item{side}{"upper" to get the upper level, "lower" to get the lower one}

\item{hierarchy}{geographic levels hierarchy name}
}
\description{
Navigate through geographic levels (get the upper level, the lower level, ...)
}
\seealso{
Other geography: 
\code{\link{geo_base_level}()},
\code{\link{geo_base_table}()},
\code{\link{geo_definition}()},
\code{\link{geo_hierarchy}()},
\code{\link{geo_is_type}()},
\code{\link{geo_level_table}()},
\code{\link{geo_level}()},
\code{\link{get_geo_join}()},
\code{\link{load_geo_zone}()},
\code{\link{load_population_age}()},
\code{\link{load_population}()}
}
\concept{geography}
