% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_level_table}
\alias{geo_level_table}
\title{Get geographic table definition for a given geographic level}
\usage{
geo_level_table(level)
}
\arguments{
\item{level}{geographic level name}
}
\description{
Get geographic table definition for a given geographic level
}
