% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_normalize.data.frame}
\alias{geo_normalize.data.frame}
\title{Normalize with geographic codes for a level after it has been loaded by the database.}
\usage{
\method{geo_normalize}{data.frame}(data, columns = NULL, levels = NULL)
}
\arguments{
\item{data}{data to normalize}

\item{columns}{columns to normalize, named vector with level name as name for each column if levels is null}

\item{levels}{corresponding level for each column. if NULL will use names of `columns` parameter}
}
\description{
Some database driver coerce column data to another type (char to int) which is can cause problems
with some code with mixed alphanumeric and numeric codes (sometimes to int, sometimes to char), cast
to int drop zerofilled codes making the code unuseable.
}
\details{
This function try to find a normalizer function associated with provided levels. Normalizer functions are
defined with the geographic levels using \code{\link{platform_geographic_levels}()}
}
