% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_normalize}
\alias{geo_normalize}
\title{Normalize with geographic codes for a level after it has been loaded by the database.}
\usage{
geo_normalize(data, ...)
}
\arguments{
\item{data}{data to normalize}

\item{...}{extra parameters for specific methods}
}
\description{
Some database driver coerce column data to another type (char to int) which is can cause problems
with some code with mixed alphanumeric and numeric codes (sometimes to int, sometimes to char), cast
to int drop zerofilled codes making the code unuseable.
}
\details{
This function try to find a normalizer function associated with provided levels. Normalizer functions are
defined with the geographic levels using \code{\link{platform_geographic_levels}()}
}
