% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_design.R
\name{get_columns_for_incidence}
\alias{get_columns_for_incidence}
\title{Returns list of columns needed to compute incidence}
\usage{
get_columns_for_incidence()
}
\description{
Returns list of columns needed to compute incidence
}
