% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_current_paths}
\alias{get_current_paths}
\title{Get the current paths defined by \code{\link{init.path}()} or \code{\link{add_path_prefix}()}}
\usage{
get_current_paths()
}
\description{
Get the current paths defined by \code{\link{init.path}()} or \code{\link{add_path_prefix}()}
}
\seealso{
Other path-functions: 
\code{\link{add_path_prefix}()},
\code{\link{init.path}()},
\code{\link{my.path}()},
\code{\link{platform_path}()},
\code{\link{print.paths_definition}()},
\code{\link{share.cache.path}()},
\code{\link{share.data.path}()},
\code{\link{share.lib.path}()},
\code{\link{share_path}()}
}
\concept{path-functions}
