% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{get_default_color}
\alias{get_default_color}
\title{Get default color}
\usage{
get_default_color(which = 1)
}
\arguments{
\item{which}{index of default color (1 or 2)}
}
\description{
Default color for a given platform can be configured in platform environment, a vector of at least 2 colors are expected
}
