% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{get_geo_hierarchy}
\alias{get_geo_hierarchy}
\title{Get a hierarchy path of geographic levels}
\usage{
get_geo_hierarchy(hierarchy = NULL)
}
\arguments{
\item{hierarchy}{name of the hierarchy path to use}
}
\description{
Depending of the context several kind of geographic levels can be used
Inside a hierarchy, all levels should be totaly nested (all sublevel should be mapped to ONE upper level)
}
\seealso{
geo.levels
}
