% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_historical_season_start}
\alias{get_historical_season_start}
\title{Get start date for each season from historical seasons definition}
\usage{
get_historical_season_start()
}
\value{
vector of start date, season number as name
}
\description{
Get start date for each season from historical seasons definition
}
