% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_historical_seasons}
\alias{get_historical_seasons}
\title{Get list of available season names}
\usage{
get_historical_seasons()
}
\value{
character vector of season names
}
\description{
Get list of available season names
}
\seealso{
\code{\link{concepts}}

Other seasons: 
\code{\link{get_historical_tables}()},
\code{\link{get_season_dates}()},
\code{\link{season_definition}()}
}
\concept{seasons}
