% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{get_historical_tables}
\alias{get_historical_tables}
\title{Get Historical tables}
\usage{
get_historical_tables()
}
\description{
Historical tables is a list, with season year (first year of the season) as name and season definition
}
\seealso{
season.def
}
