% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{get_labels_from_binary}
\alias{get_labels_from_binary}
\title{Get corresponding label from a binary group value}
\usage{
get_labels_from_binary(data, mask, group = "g", trans = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{mask}{integer vector binary mask for individual labels}

\item{group}{name of the column in data containing binary group value}

\item{trans}{translate name if TRUE use \code{\link{i18n}}, if function use it a translator}
}
\description{
Get corresponding label from a binary group value
}
