% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_season_dates}
\alias{get_season_dates}
\title{Get dates of the season}
\usage{
get_season_dates(season)
}
\arguments{
\item{season}{int season to get dates from}
}
\description{
Get dates of the season
}
\seealso{
Other seasons: 
\code{\link{get_historical_seasons}()},
\code{\link{get_historical_tables}()},
\code{\link{season_definition}()}
}
\concept{seasons}
