% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{get_symptoms_aliases}
\alias{get_symptoms_aliases}
\title{aliases for symptom columns
Aliases are meaningfull names for each columns of the survey data. InfluenzaNet surveys data are named using simple names (Q + number)
but these names are not errorproof and are hard to memorize.
epiwork.tables list defined in platform config file should describe the mapping between column names in the DB (Qxxx) and the column
names used in R.
They should not vary from one season to another (another column means another name)}
\usage{
get_symptoms_aliases()
}
\description{
aliases for symptom columns
Aliases are meaningfull names for each columns of the survey data. InfluenzaNet surveys data are named using simple names (Q + number)
but these names are not errorproof and are hard to memorize.
epiwork.tables list defined in platform config file should describe the mapping between column names in the DB (Qxxx) and the column
names used in R.
They should not vary from one season to another (another column means another name)
}
