% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{gg_barplot_percent}
\alias{gg_barplot_percent}
\title{Percentage barplot}
\usage{
gg_barplot_percent(
  values,
  order = "desc",
  label.size = rel(0.9),
  axis.label = F,
  color = NULL,
  scale_color = NULL,
  x.rotate = 90,
  x.vjust = NULL,
  label.nudge = 0.3
)
}
\arguments{
\item{values}{vector}

\item{order}{if "desc" use decreasing order}

\item{label.size}{size of the labels (can use relative size using ggplot2::rel())}

\item{axis.label}{if TRUE show axis label}

\item{color}{fill color default will use color}

\item{scale_color}{use this scale for fill colors, TRUE will use brewer palette}

\item{x.rotate}{rotation angle for labels}

\item{x.vjust}{vertical justification for x axis}

\item{label.nudge}{nudge for y-axis labels}
}
\description{
Percentage barplot
}
