% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{graph.close}
\alias{graph.close}
\title{End a graphic export to a file. Use it instead of dev.off().
this function also handle post export hooks}
\usage{
graph.close()
}
\description{
End a graphic export to a file. Use it instead of dev.off().
this function also handle post export hooks
}
