% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{graph.colors}
\alias{graph.colors}
\title{Return a list of n colors by the "best" color generator available}
\usage{
graph.colors(n, pal = NULL)
}
\arguments{
\item{n}{number of colors to get (caution, it can be fewer than expected for some palettes)}

\item{pal}{name of the palette to use}
}
\description{
Try to use RColorBrewer if available, if not uses rainbow
}
