% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{graph.open}
\alias{graph.open}
\title{Open a graph using standard library}
\usage{
graph.open(
  file,
  width = NA,
  height = NA,
  type = "png",
  pointsize = 12,
  pitch = NULL,
  ...
)
}
\arguments{
\item{file}{filename (without extension)}

\item{width}{width (pixel)}

\item{height}{height in px}

\item{type}{type of output to use}

\item{pointsize}{pointsize value}

\item{pitch}{old deprecated parameters (will raise an error if not null)}

\item{...}{extra parameted}
}
\description{
This function call the device function and apply some defaults params
}
\details{
It's useasble for classical plot functions (not ggplot2 where you can use ggsave())
}
