% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{graph.save}
\alias{graph.save}
\title{Export a ggplot2 and run graph hooks. To be used with ggplot2 graph instead of graph.close()}
\usage{
graph.save(filename, type = NULL, ...)
}
\arguments{
\item{filename}{filename of the graph}

\item{...}{others arguments passed to ggsave()}
}
\description{
Export a ggplot2 and run graph hooks. To be used with ggplot2 graph instead of graph.close()
}
