% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.r
\name{i18n.inflexion}
\alias{i18n.inflexion}
\title{Inflexion for a count
return translated string according to count}
\usage{
\method{i18n}{inflexion}(count, single, plural)
}
\arguments{
\item{count}{int count}

\item{single}{single text id if count <= 1}

\item{plural}{plural text id if count > 1}
}
\description{
Inflexion for a count
return translated string according to count
}
