% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.r
\name{i18n}
\alias{i18n}
\alias{i18n.default}
\alias{i18n.factor}
\title{Translate a string}
\usage{
i18n(x)

\method{i18n}{default}(x)

\method{i18n}{factor}(x)
}
\arguments{
\item{x}{object to translate}
}
\value{
translation or the string itself it not found
}
\description{
Translate a string
}
\examples{
i18n("my_string_to_translate")

}
