% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.r
\name{i18n_check_duplicates}
\alias{i18n_check_duplicates}
\title{Check for duplicate entry in names of the provided list
Only keep the last entry in case of duplicate}
\usage{
i18n_check_duplicates(values, from = NULL)
}
\arguments{
\item{values}{named list of entries}

\item{from}{text to use}
}
\description{
Check for duplicate entry in names of the provided list
Only keep the last entry in case of duplicate
}
