% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.R
\docType{package}
\name{ifnBase-package}
\alias{ifnBase}
\alias{ifnBase-package}
\title{InfluenzaNet Data Analysis Base Package}
\description{
This package provides base set of function to help & standardize analysis of an
InfluezaNet database.
It can be used to produce online data analysis or for research projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cturbelin/ifnBase}
  \item Report bugs at \url{https://github.com/cturbelin/ifnBase/issues}
}

}
\author{
\strong{Maintainer}: Clément Turbelin \email{clement.turbelin@sorbonne-universite.fr}

}
