% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_rs_2014_daily.R
\name{IncidenceDailyRS2014}
\alias{IncidenceDailyRS2014}
\title{Incidence estimator using rs2014 method implemented using daily agregatation}
\description{
Incidence estimator using rs2014 method implemented using daily agregatation

Incidence estimator using rs2014 method implemented using daily agregatation
}
\details{
Input:
Input data are expected to have some columns & some fixes. They are loaded and prepared using \code{\link{load_results_for_incidence}}

Parameters:

\describe{
 \item{ignore.first.delay}{Number of days form his first survey befotre to include a participant (not compatible with ignore.first)}
 \item{ignore.first.only.new}{Ignore first rule only for new participant (uses intake$first.season column}
 \item{exclude.same}{should exclude syndrom if same episode is on (partcipant still active)}
 \item{exclude.same.delay}{max delay to exclude syndrom if same (not excluded if over this delay)}
 \item{active.max.freq}{max delay between 2 surveys (in weeks)}
 \item{active.day.before}{Number of week before each computed week to count active participants}
 \item{active.day.after}{Number of week after each computed week to count active participants}
 \item{active.min.surveys}{min number of surveys for each participant (not active if less)}
 \item{active.use.mean}{Use mean number of active participant over the windows [active.day.before, active.day.after]}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{weekly}}{weekly data (loaded using \code{\link{load_results_for_incidence}})}

\item{\code{intake}}{intake data (loaded using \code{\link{load_results_for_incidence}})}

\item{\code{participant}}{data.frame() with all available participants and commputed criterias used during computation}

\item{\code{params}}{parameters for computation}

\item{\code{syndromes}}{character vector of column names containing syndromes classification for each weekly}

\item{\code{profiler}}{profiler}

\item{\code{verbose}}{logical show verbose message}

\item{\code{design}}{design stratification from \code{\link{design_incidence}}}

\item{\code{output}}{vector of character, see \code{\link{IncidenceRS2014}}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IncidenceDailyRS2014$new()}}
\item \href{#method-prepare}{\code{IncidenceDailyRS2014$prepare()}}
\item \href{#method-compute_date}{\code{IncidenceDailyRS2014$compute_date()}}
\item \href{#method-compute}{\code{IncidenceDailyRS2014$compute()}}
\item \href{#method-clone}{\code{IncidenceDailyRS2014$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
instanciate object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceDailyRS2014$new(
  weekly,
  intake,
  syndromes,
  params,
  design,
  output = c("inc")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weekly}}{data}

\item{\code{intake}}{intake data}

\item{\code{syndromes}}{syndromes column names in weekly}

\item{\code{params}}{list of parameters for computing}

\item{\code{design}}{design_incidence}

\item{\code{output}}{list of output types to compute}

\item{\code{syndroms}}{compatibility}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare"></a>}}
\subsection{Method \code{prepare()}}{
Prepare data before to compute
Prepare participants data to be able to apply selection rules
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceDailyRS2014$prepare()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_date"></a>}}
\subsection{Method \code{compute_date()}}{
Internal function
Estimate incidence with rs2014 method for a given date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceDailyRS2014$compute_date(date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{date to compute}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute"></a>}}
\subsection{Method \code{compute()}}{
Compute weekly incidence for all requested weeks
Apply the full algorithm for rs2014 incidence computation
prepare data and then compute for each week by calling estimate_incidence_rs2014_week
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceDailyRS2014$compute(dates = NULL, verbose = T, progress = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dates}}{dates to compute, if NULL will compute on all dates available in weekly data}

\item{\code{verbose}}{show verbose messages if TRUE}

\item{\code{progress}}{show progress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceDailyRS2014$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
