% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_rs_2014.R
\name{IncidenceRS2014}
\alias{IncidenceRS2014}
\title{Compute Incidence using rs2014 method
Active participants are selected using a set of rules}
\description{
Compute Incidence using rs2014 method
Active participants are selected using a set of rules

Compute Incidence using rs2014 method
Active participants are selected using a set of rules
}
\details{
Parameters:
\describe{
 \item{ignore.first.delay}{Number of days form his first survey befotre to include a participant (not compatible with ignore.first)}
 \item{ignore.first.only.new}{Ignore first rule only for new participant (uses intake$first.season column @see load_results_for_incidence)}
 \item{exclude.same}{should exclude syndrom if same episode is on (partcipant still active)}
 \item{exclude.same.delay}{max delay to exclude syndrom if same (not excluded if over this delay)}
 \item{active.max.freq}{max delay between 2 surveys (in weeks)}
 \item{active.week.before}{Number of week before each computed week to count active participants}
 \item{active.week.after}{Number of week after each computed week to count active participants}
 \item{active.min.surveys}{min number of surveys for each participant (not active if less)}
}

Input:
Input data are expected to have some columns & some fixes. They are loaded and prepared using \code{\link{load_results_for_incidence}}

Output:
Incidence estimator can output several kind of datasets
\describe{
 \item{inc}{inc=incidence at (national level)}
 \item{zlow}{"lower geographic level (z)", if estimator is stratified by geographic level}
 \item{age}{"age" age-specific incidence, age categories should be in "age.cat" column in intake data}
 \item{count}{raw count of syndromes used for each syndrom by week, before incidence is computed}
}

Apply the full algorithm for rs2014 incidence computation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{weekly}}{Weekly data}

\item{\code{intake}}{Intake data, used to build strata}

\item{\code{participant}}{A data.frame() with all available participants and commputed criterias used during computation}

\item{\code{params}}{Parameters for computation}

\item{\code{syndromes}}{Character vector of column names containing syndromes classification for each weekly}

\item{\code{profiler}}{Time Profiler}

\item{\code{verbose}}{logical produce verbose output if TRUE}

\item{\code{design}}{design stratification from \code{\link{design_incidence}}}

\item{\code{output}}{vector of character (see Details outupus)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IncidenceRS2014-new}{\code{IncidenceRS2014$new()}}
\item \href{#method-IncidenceRS2014-prepare}{\code{IncidenceRS2014$prepare()}}
\item \href{#method-IncidenceRS2014-compute_week}{\code{IncidenceRS2014$compute_week()}}
\item \href{#method-IncidenceRS2014-compute}{\code{IncidenceRS2014$compute()}}
\item \href{#method-IncidenceRS2014-clone}{\code{IncidenceRS2014$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidenceRS2014-new"></a>}}
\if{latex}{\out{\hypertarget{method-IncidenceRS2014-new}{}}}
\subsection{Method \code{new()}}{
instanciate object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceRS2014$new(
  weekly,
  intake,
  params,
  syndromes,
  design,
  output = c("inc", "zlow", "age"),
  syndroms = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weekly}}{data}

\item{\code{intake}}{intake data}

\item{\code{params}}{list of parameters for computing}

\item{\code{syndromes}}{syndromes column names in weekly}

\item{\code{design}}{design_incidence}

\item{\code{output}}{list of output types to compute}

\item{\code{syndroms}}{compatibility}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidenceRS2014-prepare"></a>}}
\if{latex}{\out{\hypertarget{method-IncidenceRS2014-prepare}{}}}
\subsection{Method \code{prepare()}}{
Prepare data for incidence computation with RS2014 method

compute first list of participants and filter with some parameters (with constant effects over time)

Internal function
Estimate incidence with rs2014 method for a given week
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceRS2014$prepare()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidenceRS2014-compute_week"></a>}}
\if{latex}{\out{\hypertarget{method-IncidenceRS2014-compute_week}{}}}
\subsection{Method \code{compute_week()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceRS2014$compute_week(yw)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{yw}}{yearweek number to compute
Compute incidence}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidenceRS2014-compute"></a>}}
\if{latex}{\out{\hypertarget{method-IncidenceRS2014-compute}{}}}
\subsection{Method \code{compute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceRS2014$compute(
  weeks = NULL,
  verbose = T,
  progress = F,
  verticalize = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weeks}}{list of weeks to compute, if NULL use all in weekly}

\item{\code{verbose}}{verbose output if TRUE}

\item{\code{progress}}{logical show progress}

\item{\code{verticalize}}{logical produce a verticalized data.frame (one column "syndrome" and type) instead of horizontal one (all syndromes as columns, one row by week)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidenceRS2014-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IncidenceRS2014-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceRS2014$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
