% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome_provider.R
\name{is_syndrome_provider}
\alias{is_syndrome_provider}
\title{Test if object is a syndrome provider object}
\usage{
is_syndrome_provider(x, accept_class = FALSE)
}
\arguments{
\item{x}{object to test}

\item{accept_class}{if TRUE class will be accepted, not only instance}
}
\description{
Test if object is a syndrome provider object
}
