% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{join_surveyuser}
\alias{join_surveyuser}
\title{Create join clause to survey_surveyuser table for a survey}
\usage{
join_surveyuser(survey_alias, user_alias)
}
\arguments{
\item{survey_alias}{table alias name of the survey table}

\item{user_alias}{table to use for the survey_user table}
}
\description{
Create join clause to survey_surveyuser table for a survey
}
