% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{load_geo_zone}
\alias{load_geo_zone}
\title{Load areas defined at a given geographic level}
\usage{
load_geo_zone(geo, type = "metro", columns = c())
}
\arguments{
\item{geo}{geographic level code (ex "nuts1")}

\item{type}{type of geographic level (if needed)}

\item{columns}{list of extra columns to fetch (if the geographic table has extra columns...)}
}
\description{
Load areas defined at a given geographic level
}
