% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{load_platform}
\alias{load_platform}
\title{Load platform file containing definitions and specific functions into .Share environment}
\usage{
load_platform()
}
\description{
Launch this function to init the package environment
It can be loaded automatically by setting 'autoload.platform' option. In this case
options should have been provided using options(ifn=list(...)) before the package is loaded see \code{\link{share.option}()}
}
\seealso{
Other platform: 
\code{\link{create_survey_definition}()},
\code{\link{platform_define_survey}()},
\code{\link{platform_import}()}
}
\concept{platform}
