% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{load_population_age}
\alias{load_population_age}
\title{Load popuplation by age-group & sex}
\usage{
load_population_age(geo, year, age.breaks = NULL, version = NULL, ...)
}
\arguments{
\item{geo}{geographic level, if null, return by age only population}

\item{year}{year of the population}

\item{age.breaks}{used to create age categories (from 5-year categories)}

\item{version}{version of the file to use as data source}

\item{...}{extra parameter}
}
\value{
data.frame('all','male','female','age.cat') for the given year
}
\description{
Load popuplation data by 5-year (or multiple) age group & sex
}
