% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_loader.R
\name{load_results_for_incidence}
\alias{load_results_for_incidence}
\title{Load data for incidence computation}
\usage{
load_results_for_incidence(season, age.categories,
  syndrome.from = list(), geo = NULL, country = NULL,
  first.season = NULL, columns = list(), onset = NULL)
}
\arguments{
\item{season}{season int[1] \code{\link{calc_season}}}

\item{age.categories}{int[], breaks to compute age groups}

\item{syndrome.from}{list() parameters to create the syndrome columns it will be used to call \code{\link{compute_weekly_syndromes}}}

\item{geo}{geographic level name to load}

\item{country}{country to load}

\item{first.season}{list of first season participants handling parameters (see details)}

\item{columns}{list() extra columns to load in each survey data (see details)}

\item{onset}{onset_design quosure expression used to compute onset date, see \code{\link{compute_onset}}, by default use \code{\link{base_onset_design}}}
}
\value{
list() with intake, weekly, syndromes (vector of name of syndrome columns)
}
\description{
This function standardize the loading of data in order to compute incidence
}
\details{
syndrome.from:
syndrome parameter will indicate how to create syndrome columns in weekly. A syndrome column is just a logical value column indicating if a weekly survey match a syndrome definition
This list will be used as arguments to call \code{\link{compute_weekly_syndromes}}

first seasons:
By default nothing is done on first season.
if first season is TRUE, then use the default of all following parameters
Otherwise provide a list with these parameters
\describe{
 \item{censored_value}{value to put to all participant when season is censored for first participation, default is FALSE}
 \item{from}{From which seasons consider the participation ("previous" or "all"), default is "all" }
}

Columns:
Several extra parameters can be provided:
\describe{
 \item{keep.all}{if TRUE keep all column in weekly data, if FALSE only keep a restricted list}
 \item{weekly}{Supplementary weekly columns to load see \code{\link{survey_load_results}}}
 \item{intake}{Supplementary intake columns to load see \code{\link{survey_load_results}}}
 \item{params}{list of parameters used as arguments but sometimes completed to an actual version, like onset if null is provided}
}

First season computing:

This feature compute a flag for each participant in the loaded intake indicating if the loaded season is the first participating season of the participant.

In european database all seasons are not available for all countries. In this case a censoring can be applied (considering all participants are not in their first season).
A function named `get_season_censoring(country)` should be available in the platform definition returing for a given country the season until wich data should
be censored (if the loaded i season is before or equal to this returned season year, the data will be censored for all participants)
}
