% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{make_week_index}
\alias{make_week_index}
\title{Make an index data.frame associating week index (wid) to a iso yearweek number}
\usage{
make_week_index(yw, col.yw = "yw", col.idx = "wid", stamp = F)
}
\arguments{
\item{yw}{int vector of yearweek number}

\item{col.yw}{chr[1] name of the yw column in the resulting data.frame}

\item{col.idx}{char[1] name of the week index column in the resulting data.frame}

\item{stamp}{logical use weekstamp instead of order of the week (depends on the range of weeks)}
}
\value{
data.frame(yw=int,wid=int)  yw=yearweek number, wid=index of the yw
}
\description{
Make an index data.frame associating week index (wid) to a iso yearweek number
}
