% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{makeWeekIndex}
\alias{makeWeekIndex}
\title{Make an index data.frame associating week index (wid) to a iso yearweek number}
\usage{
makeWeekIndex(yw, col.yw = "yw", col.idx = "wid", stamp = F)
}
\arguments{
\item{int}{yw vector of yearweek number}

\item{char[1]}{col.yw name of the yw column in the resulting data.frame}

\item{char[1]}{col.idx name of the week index column in the resulting data.frame}

\item{logical}{use weekstamp instead of order of the week (depends on the range of weeks)}
}
\value{
data.frame(yw=int,wid=int)  yw=yearweek number, wid=index of the yw
}
\description{
Make an index data.frame associating week index (wid) to a iso yearweek number
}
