% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{max_year_available}
\alias{max_year_available}
\title{Allow the population year to be the last available}
\usage{
max_year_available(year)
}
\description{
Allow the population year to be the last available
}
