% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{monday_of_week}
\alias{monday_of_week}
\title{Date of the monday of a yearweek}
\usage{
monday_of_week(yw)
}
\arguments{
\item{yw}{integer yearweek number}
}
\value{
Date
}
\description{
Date of the monday of a yearweek
}
