% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{monday_of_year}
\alias{monday_of_year}
\title{Date of the first monday of the year}
\usage{
monday_of_year(year)
}
\arguments{
\item{year}{int year}
}
\value{
Date
}
\description{
Date of the first monday of the year
}
