% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{msg}
\alias{msg}
\title{Wrapper of message() for output
You car redirect by using sink() function}
\usage{
msg(..., ln = TRUE, level = 0)
}
\description{
Wrapper of message() for output
You car redirect by using sink() function
}
