% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{multiple_xfreq}
\alias{multiple_xfreq}
\title{Frequency for multiple options by group}
\usage{
multiple_xfreq(.data, vv, group)
}
\arguments{
\item{.data}{data.frame}

\item{vv}{list of multiple options column name}

\item{group}{group for aggregation for each option}
}
\value{
list(freq, prop)
}
\description{
Frequency for multiple options by group
}
