% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{my.path}
\alias{my.path}
\title{Return the path of a file in the current ouput path}
\usage{
my.path(...)
}
\arguments{
\item{...}{characters string to used (will be concatenated)}
}
\description{
Return the path of a file in the current ouput path
}
\seealso{
Other path functions: \code{\link{init.path}}
}
