% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptoms.R
\docType{data}
\name{NO_SYMPTOM}
\alias{NO_SYMPTOM}
\title{Symptoms}
\format{
An object of class \code{character} of length 1.
}
\usage{
NO_SYMPTOM
}
\description{
No symptome variable name
}
\details{
To be used as a constant in the analysis
}
\keyword{datasets}
