% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_design.R
\name{output_incidence}
\alias{output_incidence}
\title{Define structure containing incidence results output
"output" parameter of estimate_incidence_* functions could be either a character vector with the name of each desired
output type (inc, zlow, age)
or a structure return by this function which allow to pass some extra parameters to the output function (@see calc_adjusted_incidence)}
\usage{
output_incidence(types, conf.int = T, adjust = T, ...)
}
\arguments{
\item{types}{types of measures computed}

\item{conf.int}{confidence interval computed}

\item{adjust}{adjusted measures}

\item{...}{data sets}
}
\description{
Define structure containing incidence results output
"output" parameter of estimate_incidence_* functions could be either a character vector with the name of each desired
output type (inc, zlow, age)
or a structure return by this function which allow to pass some extra parameters to the output function (@see calc_adjusted_incidence)
}
