% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{parse_season}
\alias{parse_season}
\title{Check if input is a season number}
\usage{
parse_season(season, accept.several = FALSE)
}
\arguments{
\item{season}{value to check (number or character)}
}
\value{
integer value of season number
}
\description{
Check if input is a season number
}
\keyword{internal}
