% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform_define_survey}
\alias{platform_define_survey}
\title{Describe a survey of a platform}
\usage{
platform_define_survey(name, survey_id, table, mapping, labels = NULL,
  codes = NULL, recodes = NULL, single.table = FALSE, geo.column = NULL,
  ...)
}
\arguments{
\item{name}{unique name of the survey}

\item{survey_id}{id of the survey in the database}

\item{table}{name of the default table storing response of the survey}

\item{mapping}{Mapping from DB names (in value) and R names for each survey variables.}

\item{labels}{list of labels}

\item{codes}{list of codes *deprecated*}

\item{recodes}{list of recoding}

\item{single.table}{boolean, TRUE if survey uses single table model}

\item{geo.column}{name of the variable (R names)}

\item{...}{extra parameters to set for the survey}
}
\description{
Describe structure of a survey and the mapping from DB names &
coding to R names & factor levels
}
\seealso{
survey_recode
}
