% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform_generate_seasons}
\alias{platform_generate_seasons}
\title{Generate historical tables for seasons until now}
\usage{
platform_generate_seasons(from, tables, year.pop = NULL, dates = NULL)
}
\arguments{
\item{from}{season number starting}

\item{tables}{list of tables}

\item{year.pop}{year of the population to use for the season, can be a fixed number of a function with season number as argument}

\item{dates}{function(season) return list(start, end) with date bound of the seasons, if null use default_season_dates()}
}
\description{
Generate historical tables for seasons until now
}
