% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform_geographic_tables}
\alias{platform_geographic_tables}
\title{Create geographic tables description}
\usage{
platform_geographic_tables(def, default.title = "title")
}
\arguments{
\item{def}{either a geo.levels structure or a list of table description for each level (name of the level as the name of each entry)}
}
\value{
list()
}
\description{
Create geographic tables description
}
\details{
\describe{
  \item{table}{name of the table to use}
  \item{column}{name of the column in the table containing the area id of the level}
  \item{title}{name of the column containing label of the area}
}
}
