% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform_import}
\alias{platform_import}
\title{Import a file in the platform environment}
\usage{
platform_import(name)
}
\arguments{
\item{name}{relative name in the platform path without .R extension}
}
\description{
Import a file in the platform environment
}
\seealso{
\link{concepts}

Other platform: 
\code{\link{create_survey_definition}()},
\code{\link{load_platform}()},
\code{\link{platform_define_survey}()}
}
\concept{platform}
