% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform_season_history}
\alias{platform_season_history}
\title{Define historical data}
\usage{
platform_season_history(season, dates, ...)
}
\arguments{
\item{season}{name of the season. By convention it is the year number of the starting year (2011 for 2011-2012,...)}

\item{dates}{list(start=, end=) starting & ending of each season, YYYY-MM-DD format}

\item{...}{tables names for each survey containing data of the season in case of by-season storage model.}
}
\description{
For each season you can describe how data are organized
}
\section{Data storage model}{

InfluezaNet data for a survey can be stored using 2 ways :
 - A single table contains data for all the seasons (if the survey didnt changed a lot)
 - A new table is created for each season for a given survey
}

\examples{
\dontrun{
# Single table model
platfrom_season_history('2011', dates=list(start="2011-11-15", end="2012-04-15"))
platfrom_season_history('2012', dates=list(start="2012-11-15", end="2013-04-20"))

# Multiple table model
platfrom_season_history('2011',
     dates=list(start="2011-11-15", end="2012-04-15"),
     weekly="pollster_results_weekly_2011"
)
platfrom_season_history('2012',
     dates=list(start="2012-11-15", end="2013-04-20"),
     weekly="pollster_results_weekly_2012"
)
}
}
\seealso{
Other platfom: 
\code{\link{platform_env}()},
\code{\link{platform_geographic_levels}()},
\code{\link{platform_geographic_tables}()},
\code{\link{platform_options}()}
}
\concept{platfom}
