% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform_set_seasons}
\alias{platform_set_seasons}
\title{Set all the seasons from a list}
\usage{
platform_set_seasons(seasons)
}
\arguments{
\item{seasons}{list must have season number as key}
}
\description{
Set all the seasons from a list
}
