% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{plot_age_pyramid}
\alias{plot_age_pyramid}
\title{Plot Age Pyramid}
\usage{
plot_age_pyramid(data, female, prop = T, w = 0.5, scales = list())
}
\arguments{
\item{data}{data.frame with (age.cat, gender, pop, prop)}

\item{female}{bool female rows}

\item{prop}{bool, if TRUE proportion are [0;1], if FALSE are percent [0;100]}

\item{w}{width factor}

\item{scales}{list of named char vector for "pop" & "cohort" names}
}
\description{
Plot Age Pyramid
}
