% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{polygon.ic}
\alias{polygon.ic}
\title{Draw confidence interval with polygon}
\usage{
polygon.ic(ii, col.x, col.up, col.low, col, ...)
}
\arguments{
\item{ii}{data.frame}

\item{col.x}{name of x column}

\item{col.up}{name of upper value column}

\item{col.low}{name of lower value column}

\item{col}{color}

\item{...}{extra parameters passed to polygon}
}
\description{
Draw confidence interval with polygon
}
