% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{print.episode_design}
\alias{print.episode_design}
\title{Print method for episode_design}
\usage{
\method{print}{episode_design}(x, ...)
}
\arguments{
\item{x}{episode_design object}

\item{...}{other arguments (ignored)}
}
\description{
Print method for episode_design
}
