% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_fusion.R
\name{print.fusion_strategy}
\alias{print.fusion_strategy}
\title{Print fusion strategy}
\usage{
\method{print}{fusion_strategy}(x, ...)
}
\arguments{
\item{x}{fusion_strategy object}

\item{...}{other params (compat print interface)}
}
\description{
Print fusion strategy
}
