% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{print.survey_definition}
\alias{print.survey_definition}
\title{print survey definition}
\usage{
\method{print}{survey_definition}(x, ...)
}
\arguments{
\item{x}{list() recode mapping}

\item{...}{extra parameters (print interface compatibility)}
}
\description{
print survey definition
}
