% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{print.survey_error}
\alias{print.survey_error}
\title{Print checks results}
\usage{
\method{print}{survey_error}(errors, ...)
}
\arguments{
\item{errors}{list() errors generated by \code{\link{create_survey_definition}}}
}
\description{
Print checks results
}
