% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{print.survey_labels}
\alias{print.survey_labels}
\title{print survey labels mapping}
\usage{
\method{print}{survey_labels}(x, ...)
}
\arguments{
\item{x}{list() variable mapping}

\item{...}{extra parameters (print interface compatibility)}
}
\description{
print survey labels mapping
}
