% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{print.survey_mapping}
\alias{print.survey_mapping}
\title{print survey variable mapping}
\usage{
\method{print}{survey_mapping}(x, ...)
}
\arguments{
\item{x}{list() variable mapping}

\item{...}{extra parameters (print interface compatibility)}
}
\description{
print survey variable mapping
}
