% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{print.survey_recode}
\alias{print.survey_recode}
\title{print survey recoding}
\usage{
\method{print}{survey_recode}(x, ...)
}
\arguments{
\item{x}{list() recode mapping}

\item{...}{extra parameters (print interface compatibility)}
}
\description{
print survey recoding
}
