% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{prop_by_weighted}
\alias{prop_by_weighted}
\title{Compute weighted proportion by group}
\usage{
prop_by_weighted(design, var, by)
}
\arguments{
\item{design}{survey design survey::svydesign()}

\item{var}{variable on which to compute frequency}

\item{by}{group variable name to compute total}
}
\description{
Compute weighted proportion by group
}
