% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\docType{data}
\name{PROTECTED_QUESTIONS}
\alias{PROTECTED_QUESTIONS}
\title{List of not usable for data variable names in survey mapping}
\format{An object of class \code{character} of length 3.}
\usage{
PROTECTED_QUESTIONS
}
\description{
List of not usable for data variable names in survey mapping
}
\keyword{datasets}
