% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{random_string}
\alias{random_string}
\title{Get a random string of size n
Warning not for cryptographic use}
\usage{
random_string(n)
}
\arguments{
\item{n}{length of string to generate}
}
\description{
Get a random string of size n
Warning not for cryptographic use
}
