% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_fusion.R
\name{raw_value}
\alias{raw_value}
\title{Create a list of values marked as "raw"}
\usage{
raw_value(...)
}
\arguments{
\item{...}{list of values to mark as raw}
}
\description{
This function can be used in \code{\link{episode_strategy}} to indicate that the list of values are not encoded values but raw values
(values as stored in the database)
}
