% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.r
\name{recode.weekly}
\alias{recode.weekly}
\title{Apply common recoding for weekly data}
\usage{
recode.weekly(weekly, health.status = T)
}
\arguments{
\item{weekly}{data.frame weekly data (return by survey_load_results)}

\item{health.status}{if TRUE try to get the health status for each row as stored in the db.}
}
\description{
Apply common recoding for weekly data
}
\seealso{
survey_load_health_status
}
