% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{recode_alias}
\alias{recode_alias}
\title{create a recoding alias}
\usage{
recode_alias(name)
}
\arguments{
\item{name}{name of recode to use}
}
\description{
The recoding will be proceeded using the name provided
}
