% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.r
\name{recode_intake}
\alias{recode_intake}
\title{Recode Intake variables to more error proof labels
by default translate labels used to recode variables}
\usage{
recode_intake(intake, translate = T, remove = T)
}
\arguments{
\item{intake}{data.frame() raw intake data}

\item{translate}{translate levels}

\item{remove}{bool unused}
}
\description{
Recode Intake variables to more error proof labels
by default translate labels used to recode variables
}
