% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{recode_var}
\alias{recode_var}
\title{Recode values with a given mapping}
\usage{
recode_var(x, mapping, translate = FALSE)
}
\arguments{
\item{x}{value to recode}

\item{mapping}{list of mapping db value to label, label as name and db value as value}

\item{translate}{if TRUE translate labels with \code{\link{i18n}}}
}
\value{
factor vector of recoded values
}
\description{
Recode values with a given mapping
}
