% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.r
\name{recode_weekly_date}
\alias{recode_weekly_date}
\title{Recodes weekly dates
Recodes weekly dates ensure sympt.start, fever.start & sympt.end are Date and restrict
data to the current date of the survey (date after submission will be replaced by NA)}
\usage{
recode_weekly_date(weekly)
}
\arguments{
\item{weekly}{weekly data}
}
\description{
Recodes weekly dates
Recodes weekly dates ensure sympt.start, fever.start & sympt.end are Date and restrict
data to the current date of the survey (date after submission will be replaced by NA)
}
