% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.r
\name{recode_ynp}
\alias{recode_ynp}
\title{recode Y/N/DNK where Y=0 to boolean (wow!)
Used to recode some InfluenzaNet variables to more error-proof coding
So TRUE is YES level, FALSE = NO level and NA missing value.}
\usage{
recode_ynp(x)
}
\arguments{
\item{x}{value to recode}
}
\description{
recode Y/N/DNK where Y=0 to boolean (wow!)
Used to recode some InfluenzaNet variables to more error-proof coding
So TRUE is YES level, FALSE = NO level and NA missing value.
}
