% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{register_survey_results_hook}
\alias{register_survey_results_hook}
\title{Register a function to be applied on the data when loading data of a survey}
\usage{
register_survey_results_hook(hook, force = FALSE)
}
\arguments{
\item{hook}{function with signature function(survey, data), returning transformed data}

\item{force}{logical if TRUE force replacement if hook already exists, if FALSE raise an error}
}
\description{
Register a function to be applied on the data when loading data of a survey
}
