% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{regroup.syndrome}
\alias{regroup.syndrome}
\title{Regroup mixed syndromes to non-specific for the 2012's influezanet syndromes set
Resulting levels are more pretty : "no.symptom", "ili", "cold", "gastro", "non.specific", "allergy"}
\usage{
regroup.syndrome(x)
}
\arguments{
\item{x}{values to recode}
}
\description{
Regroup mixed syndromes to non-specific for the 2012's influezanet syndromes set
Resulting levels are more pretty : "no.symptom", "ili", "cold", "gastro", "non.specific", "allergy"
}
