% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{relative_seasons}
\alias{relative_seasons}
\title{Get relative season numbers to a reference season number}
\usage{
relative_seasons(season, from = NULL)
}
\arguments{
\item{season}{int reference season}

\item{from}{int vector of relative index to the reference season}
}
\description{
Get relative season numbers to a reference season number
}
\keyword{internal}
