% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{replace.names}
\alias{replace.names}
\title{Replace names with given pairlist where value are old names and key are new names}
\usage{
replace.names(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{named list of column to rename (names are new names to assign)}
}
\description{
replace.names(data.frame(), new.name=old.name,...)
ex replace.names(d, toto="x") replace column "x" by "toto"
}
