% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{scale_x_wid}
\alias{scale_x_wid}
\title{Scale from wid to Yearweek values}
\usage{
scale_x_wid(
  ww,
  week.sep = "s",
  breaks = "auto",
  col.yw = "yw",
  label = "yw",
  by = NULL,
  week = NULL,
  complete = FALSE,
  n.pretty = 5
)
}
\arguments{
\item{ww}{data.frame() associating wid to yearweek}

\item{week.sep}{separator for yearweek week label}

\item{breaks}{start|auto|all|by|year|week breaks strategy or a function(limits) returing breaks}

\item{col.yw}{yearweek values column in [ww]}

\item{label}{labelling strategy (yw or "year" for year label), or function(yw)}

\item{by}{number of week step to use if breaks="by"}

\item{week}{number to use if breaks=week}

\item{complete}{force the ww data.frame to be complete (recompute wid) from yearweek range}

\item{n.pretty}{number ticks to generate (breaks=auto)}
}
\description{
Yearweek values are generated by \code{iso_yearweek()} function
}
\details{
start, year : tick is first week of the year
}
