% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{segment.sd}
\alias{segment.sd}
\title{Draw segments to represent a confidence Interval}
\usage{
segment.sd(i, estim, up, low, k = 0.25, col = "grey", lty = 1,
  lwd = 1, cex = 1, pch = 16)
}
\arguments{
\item{i}{x values}

\item{estim}{estimation point}

\item{up}{upper bound}

\item{low}{lower bound}

\item{k}{size of the horizontal segment}

\item{col}{color of the segments}

\item{lty}{linetype (@seealso points)}

\item{lwd}{line width}

\item{cex}{cex of the estimation point}

\item{pch}{for the estimation point}
}
\description{
Draw segments to represent a confidence Interval
}
