% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{select_df}
\alias{select_df}
\title{Select column from a data.frame,}
\usage{
select_df(data, columns, error = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{columns}{list of column to extract}
}
\value{
data.frame
}
\description{
Select column from a data.frame,
}
