% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{share.data.path}
\alias{share.data.path}
\title{path of a file in data}
\usage{
share.data.path(file = "")
}
\arguments{
\item{file}{file name to load}
}
\description{
path of a file in data
}
\seealso{
Other path-functions: 
\code{\link{add_path_prefix}()},
\code{\link{get_current_paths}()},
\code{\link{init.path}()},
\code{\link{my.path}()},
\code{\link{platform_path}()},
\code{\link{print.paths_definition}()},
\code{\link{share.cache.path}()},
\code{\link{share.lib.path}()},
\code{\link{share_path}()}
}
\concept{path-functions}
