% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{share.lib.path}
\alias{share.lib.path}
\title{Get file in lib path}
\usage{
share.lib.path(file)
}
\arguments{
\item{file}{file name to load}
}
\description{
Get file in lib path
}
\seealso{
Other path-functions: 
\code{\link{add_path_prefix}()},
\code{\link{get_current_paths}()},
\code{\link{init.path}()},
\code{\link{my.path}()},
\code{\link{platform_path}()},
\code{\link{print.paths_definition}()},
\code{\link{share.cache.path}()},
\code{\link{share.data.path}()},
\code{\link{share_path}()}
}
\concept{path-functions}
