% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.r
\name{share.lib}
\alias{share.lib}
\title{Load a share library}
\usage{
share.lib(file, force = F, platform = F, optional = F)
}
\arguments{
\item{file}{character vector of name(s) of the library to load (name of the .r file, without extension)}

\item{force}{bool, if TRUE force the loading of the lib, regardless of the cache}

\item{platform, }{look for a library in share/platform/[libname].[platform].R}

\item{optional}{if TRUE dont raise an error if the library doesnt exist. By default an error is thrown.}
}
\description{
This function is only usefull if the project comply to the project organization described in vignette #' \code{vignette("workspace", package = "ifnBase")}
}
