% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.R
\name{share.option}
\alias{share.option}
\title{get options for the ifn environment}
\usage{
share.option(...)
}
\arguments{
\item{...}{list of values to get or set}
}
\value{
list()
}
\description{
Options for the package are stored in an options() entry named 'ifn'. It should be a list
}
\details{
Packages options describe the running package environment

\describe{
 \item{graph}{ named list with default graph options list(width, height, type)}
 \item{platform}{unique id of the platform, ISO name of the country and 'eu' for european}
 \item{platform.path}{path to platform definition files}
 \item{autoload.platform}{if TRUE launch load_platform() when package is loaded}
 \item{DB_DRIVER}{name of the DB driver to use. Actually 'RPostgreSQL' or 'RODBC'}
 \item{DB_DSN}{Database connexion string (or list). for ROBDC list with 'file' entry will load DSN from the file}
 \item{base.out.path}{output path used by default init.path function}
}
}
