% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_fusion.R
\name{strategy_create_coding}
\alias{strategy_create_coding}
\title{Create a encoding dictionnary for each variable used in a strategy
Lookup for each variable to get raw values (database encoded value) and data recoding available in survey (\code{\link{survey_definition}}).
Performs some checks that values should be known and unique.}
\usage{
strategy_create_coding(data, survey = "weekly")
}
\arguments{
\item{data}{list() describing values for each variable (name as key)}

\item{survey}{survey name to use to compare variable definition (by default "weekly" survey)}
}
\value{
list() each variables will have recoded (codes) and non encoded (raw) values
}
\description{
Create a encoding dictionnary for each variable used in a strategy
Lookup for each variable to get raw values (database encoded value) and data recoding available in survey (\code{\link{survey_definition}}).
Performs some checks that values should be known and unique.
}
