% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_aliases}
\alias{survey_aliases}
\title{Convert DB column name from and to question variable name (aliases)}
\usage{
survey_aliases(cols, def, revert = FALSE)
}
\arguments{
\item{cols}{names of columns to rename}

\item{def}{list of aliases}

\item{revert}{if TRUE, aliases to Db names, if false (default) DB names to aliases}
}
\description{
Description of mapping are defined in the survey  by \code{\link{platform_define_survey}} in the platform file
}
\seealso{
Other survey: 
\code{\link{survey_definition}()},
\code{\link{survey_labels}()},
\code{\link{survey_recode}()},
\code{\link{survey_recodings}()},
\code{\link{survey_rename_columns}()},
\code{\link{survey_variable_recoding}()},
\code{\link{survey_variables_like}()}
}
\concept{survey}
