% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_default_language}
\alias{survey_default_language}
\title{Get the default language used for survey
Uses 'survey_default_language' variable that should be defined in platform file}
\usage{
survey_default_language()
}
\description{
Get the default language used for survey
Uses 'survey_default_language' variable that should be defined in platform file
}
