% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_labeller.R
\name{survey_labeller}
\alias{survey_labeller}
\title{Get a labeller object for a given name in the survey
Will to replace \code{\link{survey_labels}()} returns a \code{\link{SurveyLabeller}} object}
\usage{
survey_labeller(survey, name)
}
\description{
Get a labeller object for a given name in the survey
Will to replace \code{\link{survey_labels}()} returns a \code{\link{SurveyLabeller}} object
}
\seealso{
Other survey: 
\code{\link{SurveyLabeller}},
\code{\link{survey_aliases}()},
\code{\link{survey_definition}()},
\code{\link{survey_labels}()},
\code{\link{survey_recode}()},
\code{\link{survey_recodings}()},
\code{\link{survey_rename_columns}()},
\code{\link{survey_variable_recoding}()},
\code{\link{survey_variables_like}()}
}
\concept{survey}
