% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_labels}
\alias{survey_labels}
\title{Return the "labels" for a given question}
\usage{
survey_labels(survey, question)
}
\arguments{
\item{survey}{name of the survey}

\item{question}{name of the question, or if multichoice name of the question group (sometimes a more generic name)}
}
\description{
If it is a multichoice question, it will return the names of columns containing all the responses choices for this question
If it is a single choice question, it will return a list of language independent labels (but human meaningful)
}
