% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.r
\name{survey_load_health_status}
\alias{survey_load_health_status}
\title{Load health status for each weekly responses}
\usage{
survey_load_health_status(weekly, health.table = NULL)
}
\arguments{
\item{weekly}{a data.frame loaded by survey_load_results (weekly survey)}

\item{health.table}{name of a health status coding view. CAUTION if used, no check is done (is right table used for the right season)}
}
\value{
weekly with extra column 'status'
}
\description{
load health status from the table (or view) in db. If weekly are store in separated tables by season,
the weekly data.frame should have a "season" attribute telling which table to use (historical.tables variable should map table for each seasons)
}
