% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_loader.R
\name{survey_load_options}
\alias{survey_load_options}
\title{load options of a question from the database, identified by its question_id
low level function}
\usage{
survey_load_options(question_id, translation_id = NULL)
}
\arguments{
\item{question_id}{int[]}

\item{translation_id}{id the translation to use}
}
\value{
data.frame() option_id, text
}
\description{
load options of a question from the database, identified by its question_id
low level function
}
\seealso{
Other survey definition functions: 
\code{\link{survey_load_questions}()}
}
\concept{survey definition functions}
