% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_loader.R
\name{survey_load_questions}
\alias{survey_load_questions}
\title{load questions of a survey from the database}
\usage{
survey_load_questions(survey, translation_id = NULL)
}
\arguments{
\item{survey}{survey name (@see epiwork.tables)}

\item{translation_id}{id of the translation to use}
}
\description{
load questions of a survey from the database
}
\seealso{
Other survey-load: 
\code{\link{survey_load_all}()},
\code{\link{survey_load_options}()},
\code{\link{survey_load_participants}()},
\code{\link{survey_load_participations}()},
\code{\link{survey_load_results_historic}()},
\code{\link{survey_load_results}()},
\code{\link{survey_load_translation}()},
\code{\link{survey_options_for}()},
\code{\link{survey_participant_previous_season}()},
\code{\link{survey_participant_season}()},
\code{\link{survey_question_id}()}
}
\concept{survey definition functions}
\concept{survey-load}
