% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{survey_load_results}
\alias{survey_load_results}
\title{load a survey results set}
\usage{
survey_load_results(survey, cols, geo = NULL, date = NULL,
  db.table = NULL, survey.users = NULL, debug = F, account = F,
  where = c(), season = NULL, channel = NULL, cols.sup = c(),
  gid = F, country = NULL)
}
\arguments{
\item{survey}{: survey name (as defined in epiwork.tables)}

\item{cols}{column list (you can use aliases)}

\item{geo}{code for the geo levels to add to the results set (only if the survey contains a geographic column)}

\item{date}{list(min=,max=) of min and max date to restrict the loaded data (use of the timestamp of the survey)}

\item{db.table}{override table name (to hanlde backup tables or test table)}

\item{survey.users}{list of id of SurveyUsers, load data only for these ids}

\item{debug}{if TRUE show query}

\item{account}{boolean if TRUE include "account" id (account user id). Only works if "user" column is defined}

\item{where}{where condition to apply on the query (caution, no check)}

\item{season}{if not NULL, season number, will load data for a given season, by default load all data from "pollster_results_[survey]"}

\item{channel}{if not null get data only for a given channel value}

\item{cols.sup}{supplementary column or SQL select clauses (caution no check)}

\item{gid}{if TRUE load gid column}

\item{country}{country name (if survey data has country column)}
}
\value{
data.frame data of the survey
}
\description{
survey_load_results loads results set for a given survey. It handles several features:
 \itemize{
   \item mapping of column names (in the DB) to meaningful names (defined in epiwork.tables)
   \item merging with geographic tables (geo parameters set geographic level to use)
   \item selecting data from a season (automatic table and/or dates bound selection)
}
}
\details{
Note : "global_id" is not used here we use person_id (sequential index of the user = primary key of survey_surveyuser table) instead because it is really cheaper for memory
}
