% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_load_results_historic}
\alias{survey_load_results_historic}
\title{Load historical data for a set of users}
\usage{
survey_load_results_historic(ids, survey, cols)
}
\arguments{
\item{ids}{list of participants (survey_user.id or results from survey_load_participants)}

\item{survey}{survey shortname}

\item{cols}{list of columns to load}
}
\description{
Load historical data for a set of users
}
