% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_load_translation}
\alias{survey_load_translation}
\title{load a translation for a survey for a given language}
\usage{
survey_load_translation(survey, language = survey_default_language())
}
\arguments{
\item{survey}{survey name}

\item{language}{code (as used in the db)}
}
\description{
load a translation for a survey for a given language
}
