% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_loader.R
\name{survey_options_for}
\alias{survey_options_for}
\title{Get the options for a survey's varname
This is not the "options" defined}
\usage{
survey_options_for(survey, varname)
}
\arguments{
\item{survey}{survey name}

\item{varname}{variable name}
}
\description{
Get the options for a survey's varname
This is not the "options" defined
}
\seealso{
Other survey-load: 
\code{\link{survey_load_all}()},
\code{\link{survey_load_options}()},
\code{\link{survey_load_participants}()},
\code{\link{survey_load_participations}()},
\code{\link{survey_load_questions}()},
\code{\link{survey_load_results}()},
\code{\link{survey_load_results_historic}()},
\code{\link{survey_load_translation}()},
\code{\link{survey_participant_previous_season}()},
\code{\link{survey_participant_season}()},
\code{\link{survey_question_id}()}
}
\concept{survey-load}
