% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{survey_participant_season}
\alias{survey_participant_season}
\title{Get list of participant id (person_id = survey_user_id)
List of participants registred in weekly at least once for a given season}
\usage{
survey_participant_season(season, use.season.dates = F)
}
\arguments{
\item{season}{season number to get}

\item{use.season.dates}{restrict to season's starting & ending dates.}
}
\description{
Get list of participant id (person_id = survey_user_id)
List of participants registred in weekly at least once for a given season
}
\seealso{
Other survey-load: 
\code{\link{survey_load_all}()},
\code{\link{survey_load_options}()},
\code{\link{survey_load_participants}()},
\code{\link{survey_load_participations}()},
\code{\link{survey_load_questions}()},
\code{\link{survey_load_results_historic}()},
\code{\link{survey_load_results}()},
\code{\link{survey_load_translation}()},
\code{\link{survey_options_for}()},
\code{\link{survey_participant_previous_season}()},
\code{\link{survey_question_id}()}
}
\concept{survey-load}
