% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_question_recoding}
\alias{survey_question_recoding}
\title{Get the recoding mapping of a survey}
\usage{
survey_question_recoding(survey, question, must.exists = TRUE)
}
\arguments{
\item{survey}{character survey name}

\item{question}{}
}
\value{
vector with label as name, database value as value
}
\description{
Get the recoding mapping of a survey
}
