% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_questions_like}
\alias{survey_questions_like}
\title{Get list of question aliases matching the given pattern (glob style)}
\usage{
survey_questions_like(survey, pattern)
}
\arguments{
\item{survey}{survey name}

\item{pattern}{glob style pattern (ex "visit.*")}
}
\value{
character vector of questions
}
\description{
These labels are defined in the survey  by \code{\link{platform_define_survey}} in the platform file
}
\examples{
\dontRun{
 survey_questions_like("weekly", "visit.*") # All variables starting with visit.*
}

}
