% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_recode}
\alias{survey_recode}
\title{Recode data from the DB storage code to the R labels (or to its translaction into default language)}
\usage{
survey_recode(x, question, survey, translate = F)
}
\arguments{
\item{x}{values to recodes}

\item{question}{name of the question (alias here)}

\item{survey}{name of the survey}

\item{translate}{if TRUE, try to translate the labels (using i18n function @seealso i18n)}
}
\description{
This function recode the code used in the DB and replace each values by a label more meaningful (like the column name but applied
on response values of a question). Mapping between DB codes and labels are described in the survey's section of epiwork.tables
(defined in platform file):
\describe{
 \item{codes}{section are the list of the DB codes}
 \item{labels}{section are the labels corresponding to each code (in the same order)}
}
}
