% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_recodings}
\alias{survey_recodings}
\title{Get all the recodings defined in a survey}
\usage{
survey_recodings(survey)
}
\arguments{
\item{survey}{character survey name}
}
\value{
list with question variable name in name, a mapping as value
}
\description{
Get all the recodings defined in a survey
}
\seealso{
Other survey: 
\code{\link{SurveyLabeller}},
\code{\link{keep_last_survey}()},
\code{\link{survey_aliases}()},
\code{\link{survey_definition}()},
\code{\link{survey_labeller}()},
\code{\link{survey_labels}()},
\code{\link{survey_recode}()},
\code{\link{survey_rename_columns}()},
\code{\link{survey_variable_recoding}()},
\code{\link{survey_variables_like}()}
}
\concept{survey}
