% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_rename_columns}
\alias{survey_rename_columns}
\title{Rename columns to the survey variable names}
\usage{
survey_rename_columns(data, survey)
}
\arguments{
\item{data}{data.frame() containing columns using database column names}

\item{survey}{character survey name}
}
\value{
list with question variable name in name, a mapping as value
}
\description{
This function can be used when loading data stored in the database format (with db column names)
}
\seealso{
Other survey: 
\code{\link{SurveyLabeller}},
\code{\link{survey_aliases}()},
\code{\link{survey_definition}()},
\code{\link{survey_labeller}()},
\code{\link{survey_labels}()},
\code{\link{survey_recode}()},
\code{\link{survey_recodings}()},
\code{\link{survey_variable_recoding}()},
\code{\link{survey_variables_like}()}
}
\concept{survey}
