% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_single_table}
\alias{survey_single_table}
\title{Returns TRUE if the survey's data are store using a single table model}
\usage{
survey_single_table(survey)
}
\arguments{
\item{survey}{survey name}
}
\description{
Returns TRUE if the survey's data are store using a single table model
}
