% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{survey_template}
\alias{survey_template}
\title{Survey Template definitions}
\usage{
survey_template(name)
}
\description{
Templates of Surveys
}
\details{
Template for common surveys : intake & weekly

Contains mapping definition and recoding for the common part of InfluenzaNet surveys

names:
\items{
 \item{eu:intake}
 \item{eu:weekly}
}
}
\seealso{
\code{\link{platform_define_survey}}
}
