% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\docType{data}
\name{survey_templates}
\alias{survey_templates}
\title{Survey Template definitions
Templates of Surveys}
\format{An object of class \code{list} of length 2.}
\usage{
survey_templates
}
\description{
Template for common surveys : intake & weekly
}
\details{
Contains mapping definition and recoding for the common part of InfluenzaNet surveys

names:
\items{
 \item{eu:intake}
 \item{eu:weekly}
}
}
\seealso{
\code{platform_define_survey}
}
\keyword{datasets}
