% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{survey_variable_available}
\alias{survey_variable_available}
\title{Check if variables are available for a given season}
\usage{
survey_variable_available(variables, survey, season)
}
\arguments{
\item{variables}{character vector of name of variables (not db name)}

\item{survey}{survey name or survey_definition from \code{\link{survey_definition}()}}

\item{season}{season to check of variable availability}
}
\value{
variables if they are available
}
\description{
This function is conservative: only variable registred with an availability clause are checked
}
