% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_variable_recoding}
\alias{survey_variable_recoding}
\title{Get the recoding mapping of a variable in a survey}
\usage{
survey_variable_recoding(survey, name, must.exists = TRUE)
}
\arguments{
\item{survey}{character survey name}

\item{name}{question variable name}
}
\value{
vector with label as name, database value as value
}
\description{
Get the recoding mapping of a variable in a survey
}
