% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_labeller.R
\name{SurveyLabeller}
\alias{SurveyLabeller}
\title{Survey Labeller Class}
\description{
Class to manipulate survey 'labels' (list of labels or columns)
Labeller object expose several methods useful to manipulate labels set
}
\seealso{
Survey section \link{concepts}

Other survey: 
\code{\link{survey_aliases}()},
\code{\link{survey_definition}()},
\code{\link{survey_labeller}()},
\code{\link{survey_labels}()},
\code{\link{survey_recode}()},
\code{\link{survey_recodings}()},
\code{\link{survey_rename_columns}()},
\code{\link{survey_variable_recoding}()},
\code{\link{survey_variables_like}()}
}
\concept{survey}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{labels}}{List of values to be used a labels}

\item{\code{glob}}{pattern if the label is implicitly defined by a pattern}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SurveyLabeller-new}{\code{SurveyLabeller$new()}}
\item \href{#method-SurveyLabeller-list}{\code{SurveyLabeller$list()}}
\item \href{#method-SurveyLabeller-unprefix}{\code{SurveyLabeller$unprefix()}}
\item \href{#method-SurveyLabeller-clone}{\code{SurveyLabeller$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyLabeller-new"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyLabeller-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyLabeller$new(survey, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{survey}}{survey name}

\item{\code{name}}{labels set name in the survey definition
Get list of labels}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyLabeller-list"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyLabeller-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyLabeller$list(.data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{only keep labels present in names of this dat
Remove common prefix labels (only works with glob pattern defined labels like 'vacc.flu.*')}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyLabeller-unprefix"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyLabeller-unprefix}{}}}
\subsection{Method \code{unprefix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyLabeller$unprefix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
function to remove prefix, can be use as a labeller function (in ggplot2 for example)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyLabeller-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyLabeller-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyLabeller$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
