% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{symptoms.ecdc}
\alias{symptoms.ecdc}
\title{Compute ECDC ILI Definition}
\usage{
symptoms.ecdc(r, with.chills = FALSE)
}
\arguments{
\item{r}{weekly data}

\item{with.chills}{use chills-like symptoms}
}
\description{
Compute ECDC ILI Definition
}
