% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{symptoms.gastro}
\alias{symptoms.gastro}
\title{Compute a generic gastroenteriritis definition (broader than French Sentinel definition)}
\usage{
symptoms.gastro(r)
}
\arguments{
\item{r}{data.frame() with syndrom columns}
}
\description{
Compute a generic gastroenteriritis definition (broader than French Sentinel definition)
}
