% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{symptoms.grog}
\alias{symptoms.grog}
\title{Compute GROG (French Influenza Surveillance historical association)'s ARI definition
Sudden syndrome with at least one general symptom and respiratory (cough or coryza or rhinitis)}
\usage{
symptoms.grog(r)
}
\arguments{
\item{r}{weekly data}
}
\description{
Compute GROG (French Influenza Surveillance historical association)'s ARI definition
Sudden syndrome with at least one general symptom and respiratory (cough or coryza or rhinitis)
}
