% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{symptoms.ili.fever}
\alias{symptoms.ili.fever}
\title{Compute ili.fever syndrome classification
Customization of ecdc}
\usage{
symptoms.ili.fever(r, level = F, include.miss = F)
}
\arguments{
\item{r}{weekly data}

\item{level}{if TRUE take fever level into account, if FALSE only uses symptom "fever"}

\item{include.miss}{if TRUE include DK and missing fever level values in "highest.temp"}
}
\description{
Compute ili.fever syndrome classification
Customization of ecdc
}
