% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{symptoms.ili.rs}
\alias{symptoms.ili.rs}
\title{Get the ILI French Sentinel network syndrome}
\usage{
symptoms.ili.rs(r, infer.fever.level = "low", pain.age.limit = NA)
}
\arguments{
\item{r}{data.frame}

\item{infer.fever.level}{= how to infer the fever level when it's missing (low = consider it's low, "high" consider it's high)}

\item{pain.age.limit, }{if defined ignore "pain" criterium for age under this limit (=consider always true for pain symptom)}
}
\description{
Get the ILI French Sentinel network syndrome
}
