% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{symptoms.influenzanet}
\alias{symptoms.influenzanet}
\title{Compute Influenzanet ILI Definition (ECDC variation)}
\usage{
symptoms.influenzanet(r)
}
\arguments{
\item{r}{weekly data}
}
\description{
Compute Influenzanet ILI Definition (ECDC variation)
}
