% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{symptoms.ira}
\alias{symptoms.ira}
\title{Acute Upper Respiratory Infection}
\usage{
symptoms.ira(r, wide.respiratory = F)
}
\arguments{
\item{r}{weekly data}

\item{wide.respiratory}{consider a wider respiratory syndrome if TRUE}
}
\description{
Acute Upper Respiratory Infection
}
