% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptoms.R
\docType{data}
\name{SYMPT_FEVER}
\alias{SYMPT_FEVER}
\alias{SYMPT_CHILLS}
\alias{SYMPT_RHINO}
\alias{SYMPT_SNEEZE}
\alias{SYMPT_SORETHROAT}
\alias{SYMPT_COUGH}
\alias{SYMPT_DYSPNEA}
\alias{SYMPT_HEADACHE}
\alias{SYMPT_PAIN}
\alias{SYMPT_CHESTPAIN}
\alias{SYMPT_ASTHENIA}
\alias{SYMPT_ANOREXIA}
\alias{SYMPT_SPUTUM}
\alias{SYMPT_WATERYEYE}
\alias{SYMPT_NAUSEA}
\alias{SYMPT_VOMITING}
\alias{SYMPT_DIARRHEA}
\alias{SYMPT_ABDOPAIN}
\alias{SYMPT_OTHER}
\alias{SYMPT_SMELL}
\alias{SYMPT_TASTE}
\alias{SYMPT_NOSEBLEED}
\alias{SYMPTOMS}
\title{Symptoms}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 22.
}
\usage{
SYMPT_FEVER

SYMPT_CHILLS

SYMPT_RHINO

SYMPT_SNEEZE

SYMPT_SORETHROAT

SYMPT_COUGH

SYMPT_DYSPNEA

SYMPT_HEADACHE

SYMPT_PAIN

SYMPT_CHESTPAIN

SYMPT_ASTHENIA

SYMPT_ANOREXIA

SYMPT_SPUTUM

SYMPT_WATERYEYE

SYMPT_NAUSEA

SYMPT_VOMITING

SYMPT_DIARRHEA

SYMPT_ABDOPAIN

SYMPT_OTHER

SYMPT_SMELL

SYMPT_TASTE

SYMPT_NOSEBLEED

SYMPTOMS
}
\description{
Registered symptoms and constants
}
\keyword{datasets}
