% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptoms.R
\docType{data}
\name{SYMPTOMS_TOPICS}
\alias{SYMPTOMS_TOPICS}
\title{Symptoms grouped by topics}
\format{
An object of class \code{list} of length 7.
}
\usage{
SYMPTOMS_TOPICS
}
\description{
Symptoms grouped by topics
}
\keyword{datasets}
