% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{syndrome.provider.rs}
\alias{syndrome.provider.rs}
\title{Compute syndromes classification using several definitions, for French incidence computation}
\usage{
syndrome.provider.rs(weekly, intake)
}
\arguments{
\item{weekly}{data.frame() of weekly results (should have the necessary columns to compute it)}

\item{intake}{intake data, (with age column)}
}
\description{
Return data.frame of French (rs) syndrome computed for each row of the weekly df
}
