% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome_provider.R
\docType{data}
\name{SyndromeProvider}
\alias{SyndromeProvider}
\title{Syndrome provider base class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SyndromeProvider
}
\description{
Syndrome provider base class
}
\keyword{datasets}
