% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome_provider_country.R
\docType{class}
\name{SyndromeProviderCountry}
\alias{SyndromeProviderCountry}
\title{Syndrome Provider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SyndromeProviderCountry
}
\description{
Syndrome Provider
}
\section{Fields}{

\describe{
\item{\code{country}}{country}

\item{\code{season}}{season number}

\item{\code{use.fever.level}}{Use fever level when definition request a fever level or only the fever=TRUE}

\item{\code{unknown.has.fever}}{if TRUE consider that unknown fever status is fever=TRUE}

\item{\code{min.syndrome.count}}{Minimum syndrom count to take into account when definition doesnt say it}

\item{\code{Age}}{to take into account of pain. Under this age pain=TRUE (disabled if NA)}
}}

\keyword{datasets}
