% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome_provider_ifn.R
\name{SyndromeProviderIfn}
\alias{SyndromeProviderIfn}
\title{Syndrome provider base class for historical Influenzanet syndromic classifier}
\description{
Syndrome provider base class for historical Influenzanet syndromic classifier

Syndrome provider base class for historical Influenzanet syndromic classifier
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SyndromeProviderIfn$new()}}
\item \href{#method-compute}{\code{SyndromeProviderIfn$compute()}}
\item \href{#method-clone}{\code{SyndromeProviderIfn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderIfn$new(version = 2012, regroup = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{classifier version to use, default is 2012}

\item{\code{regroup}}{use grouped version of syndromes for 2012 if TRUE (default)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute"></a>}}
\subsection{Method \code{compute()}}{
Compute syndrome
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderIfn$compute(weekly, intake)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weekly}}{weekly data}

\item{\code{intake}}{intake data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame() with id column from weekly and extra column for each syndrome
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderIfn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
