% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndromes_provider_2019.R
\docType{data}
\name{SyndromeProviderRS2019}
\alias{SyndromeProviderRS2019}
\title{Syndrome Provider with 2019 revised ILI & Ari definitions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SyndromeProviderRS2019
}
\description{
To run need weekly with get_columns_for_incidence() column list
}
\examples{
\dontrun{
 season = 2017
 weekly = survey_load_results("weekly", get_columns_for_incidence(), season=season)
 intake = survey_load_results("weekly", c('timestamp','date.birth'))
 weekly = recode_weekly(weekly)
 intake = recode_intake(intake)
 provider = SyndromeProviderRS2019$new()
 r = provider$compute(weekly, intake)
}

}
\keyword{datasets}
