% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{syndromes.influenzanet.2011}
\alias{syndromes.influenzanet.2011}
\title{Compute the 2011's Influenzanet syndromes set  (NO-SYMPT/ILI/COMMON-COLD/GASTRO/OTHER)}
\usage{
syndromes.influenzanet.2011(r, as.levels = F)
}
\arguments{
\item{r}{data.frame of weekly results}
}
\description{
Compute the 2011's Influenzanet syndromes set  (NO-SYMPT/ILI/COMMON-COLD/GASTRO/OTHER)
}
\seealso{
syndromes.influenzanet.2012
}
