% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\docType{data}
\name{syndromes.set}
\alias{syndromes.set}
\title{Syndromes set definitions
Defines each known syndrome set, with possible levels, pretty levels (regrouped version)
also define a provider function, usable to compute the syndrome classification from weekly data}
\format{An object of class \code{list} of length 3.}
\usage{
syndromes.set
}
\description{
Syndromes set definitions
Defines each known syndrome set, with possible levels, pretty levels (regrouped version)
also define a provider function, usable to compute the syndrome classification from weekly data
}
\keyword{datasets}
