% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{syndromes_influenzanet_2011}
\alias{syndromes_influenzanet_2011}
\title{Compute the 2011's Influenzanet syndromes set  (NO-SYMPT/ILI/COMMON-COLD/GASTRO/OTHER)}
\usage{
syndromes_influenzanet_2011(r, as.levels = F)
}
\arguments{
\item{r}{data.frame of weekly results}

\item{as.levels}{return levels instead of numeric values}
}
\description{
Compute the 2011's Influenzanet syndromes set  (NO-SYMPT/ILI/COMMON-COLD/GASTRO/OTHER)
}
\seealso{
syndromes.influenzanet.2012
}
