% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndromes_rs_2018.R
\name{syndromes_rs_2018}
\alias{syndromes_rs_2018}
\title{Compute syndromes definitions used in Guerrisi C, Euro Surv. 2018.}
\usage{
syndromes_rs_2018()
}
\value{
list() list of functions
}
\description{
This function embeds in one function to reduce number of available functions since it's provided for archive purpose
This function are deprecated.
}
\details{
\describe{
  \item{compute(weekly, intake)}{returns a data.frame() with a column for each syndrome definition and "id" as weekly survey id}
}
}
