% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.r
\name{titlelize}
\alias{titlelize}
\title{Create a title from a set of character vectors
Each argument will be translated and then concatenated in a single string
Caution, this function collapse all arguments, and return ONE string}
\usage{
titlelize(..., translate = T)
}
\arguments{
\item{...}{list of character string fragments to titlelize}

\item{translate}{bool use i18n for each fragment}
}
\description{
Create a title from a set of character vectors
Each argument will be translated and then concatenated in a single string
Caution, this function collapse all arguments, and return ONE string
}
