% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_plot}
\alias{upset_plot}
\title{Create an upset graph, inspired from an UpsetR package, show associations}
\usage{
upset_plot(
  data,
  sets,
  n.max = 40,
  point_size = 3,
  name_size_scale = 1,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  opts = list()
)
}
\arguments{
\item{data}{data.frame() with data for each sets}

\item{sets}{list of column names for each variable in sets}

\item{n.max}{number of top obs to keep, (all if na)}

\item{point_size}{size of point in intersection plot}

\item{name_size_scale}{scale of labels size in intersection plot}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}

\item{caption}{caption of the plot}

\item{opts}{options for freq plot}
}
\description{
re-implementation of the graph to separate association computation and vizualization
}
\section{Options}{


General options :
- theme: add the theme to the plot
- freq: frequency plot option
- matrix: matrix plot option
}

\section{Frequency plot options}{

\describe{
  \item{bar.color}{}
  \item{bar.fill}{fill color}
  \item{label}{label of the axis}
  \item{prop}{list, Compute cumulative proportion options}
}
}

\section{Matrix plot options}{

\describe{
  \item{point.color}{Color of point}
  \item{point.empty}{Color if set size is empty}
  \item{point.size}{Size of the point}
  \item{name.size}{Size of the name}
  \item{grid.color}{Color of the grid}
  \item{line.size}{Size of the line connecting association}
  \item{line.color}{color of the line connecting association}
  \item{set.labels}{vector of labels of each set name or function(lab) to get set label from set name}
  \item{order.freq}{reorder set list by frequency}
 }
}

\seealso{
Other upset: 
\code{\link{apply_binary_mask}()},
\code{\link{create_binary_groups}()},
\code{\link{create_binary_mask}()},
\code{\link{get_labels_from_binary}()}
}
\concept{upset}
