% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{validate_platform}
\alias{validate_platform}
\title{Post loading function to validate platform info}
\usage{
validate_platform()
}
\description{
Post loading function to validate platform info
}
\keyword{internal}
