% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_utils.R
\name{var_labels}
\alias{var_labels}
\title{Define a label for survey variables, labels are used to identify a subset of variables using a name}
\usage{
var_labels(labels, exclude = NULL)
}
\arguments{
\item{labels}{list of label or glob pattern to identify variables matching the list}

\item{exclude}{list of names to exclude after creating the list with a pattern}
}
\value{
character
}
\description{
The name is useable with \code{\link{survey_labels}} to get the list of the variables
}
\details{
A label can be defined as a list of variables names or as a glob pattern (like 'myvar.*') if variables
are named using a convenient naming convention.
}
