% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_utils.R
\name{variable_available}
\alias{variable_available}
\title{Define availability of variables}
\usage{
variable_available(value, seasons)
}
\arguments{
\item{value}{database name}

\item{seasons}{either a vector of season or an rlang::quosure (see details)}
}
\description{
Define availability of variables
}
\details{
If seasons is a quosure it will be evaluated in the survey_load_results environment.
It must be evaluate as a single logical result
Especially the expression can evaluate the season
}
\seealso{
Other platform-helpers: 
\code{\link{override}()},
\code{\link{recode_alias}()},
\code{\link{var_labels}()}
}
\concept{platform-helpers}
