% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_utils.R
\name{verticalize_incidence}
\alias{verticalize_incidence}
\title{Verticalize incidence data after computation}
\usage{
verticalize_incidence(inc, ids, syndroms)
}
\description{
Incidence data are organized horizontally, estimation for each syndroms are in columns (with suffixes .upper, .lower, .w2)
}
\seealso{
\code{\link{calc_adjusted_incidence}}
}
