% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{week_of_date}
\alias{week_of_date}
\title{week_of_date(from,to)
get yearweek from a date interval date interval}
\usage{
week_of_date(from, to)
}
\arguments{
\item{from}{date (Date class)}

\item{to}{Date class}
}
\value{
dataframe(dates, yearweek number, using the format YYYYWW, so Year * 100 + Week in the year)
}
\description{
week_of_date(from,to)
get yearweek from a date interval date interval
}
