% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{weeksbydates}
\alias{weeksbydates}
\title{weeksbydates(from,to)
get yearweeks of each date in a date interval}
\usage{
weeksbydates(from, to)
}
\arguments{
\item{from}{date (Date class)}

\item{to}{Date class}
}
\value{
dataframe(dates, yearweek number, using the format YYYYWW, so Year * 100 + Week in the year)
}
\description{
weeksbydates(from,to)
get yearweeks of each date in a date interval
}
