% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{WeekStamp}
\alias{WeekStamp}
\title{Compute a "Week" stamp (like timestamp) : week number from 1970-01-01}
\usage{
WeekStamp(yw)
}
\arguments{
\item{yw}{yearweek value}
}
\description{
This provide a continuous index for week number which not depend on period bound (unlike makeWeekIndex)
It is usefull to plot week based data.
}
