% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{with_ggplot}
\alias{with_ggplot}
\title{Init ggplot2 base theme. ggplot2 library should be loaded before !}
\usage{
with_ggplot(theme = NULL)
}
\arguments{
\item{theme}{theme() value to use as default theme}
}
\description{
Recall it with another theme
}
