% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{YearStart}
\alias{YearStart}
\title{Date of the first monday of the Year}
\usage{
YearStart(WhichYear)
}
\arguments{
\item{WhichYear}{int year}
}
\value{
Date
}
\description{
Date of the first monday of the Year
}
