% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_design.R
\name{calc_adjusted_incidence}
\alias{calc_adjusted_incidence}
\title{Compute crude & adjusted incidence from a count data.frame
This function is shared by all incidence computation methods, once active participants & syndrom counts are provided}
\usage{
calc_adjusted_incidence(count.week, design, syndroms, output)
}
\arguments{
\item{count.week}{data.frame(), count for a week. Count of active participants is in "active" column}

\item{design}{design structure (defining strata and reference data like population in each strata)}

\item{syndroms}{char() list of column names for each syndrom count}

\item{output}{char() list of requested output "inc"=global incidence, "zlow"=lower geo level incidence, "age" by age incidence,}
}
\description{
Ouput columns are ([syndrom] is the column with logical value for presence of one syndrom in the weekly data, for example "ili"):
 - active : count of active participants
 - [syndrom] : count for this syndrom (or variable)
 - [syndrom].crude : crude incidence (without adjustment, count/active participant ratio)
 - [syndrom].adj : Adjusted incidence (taking into account of stratification design)
 - [syndrom].crude.(upper|lower) : upper & lower bound of CI95 for crude incidence (Exact poisson CI)
 - [syndrom].adj.(upper|lower) : upper & lower boud of CI95 for ajusted incidence (DKES method)
}
