% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{check_list_mapping}
\alias{check_list_mapping}
\title{Check if two named list are compatible to merge}
\usage{
check_list_mapping(new, old, raise, only.errors = TRUE)
}
\arguments{
\item{new}{list()}

\item{old}{list()}

\item{raise}{function(type, value, problem, message)}

\item{only.errors}{only raise errors}
}
\description{
To be compatible with template, mapping should follow these rules:
\itemize{
  \item{only one value is mapped to a name}
  \item{new can add new name with new value, not already in old}
  \item{new cannot redefine a value with a new name (unless allow_override, not currently supported)}
}
}
\details{
The checks are transmitted to a provided function raise()
}
