% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_loader.R
\name{complete_intake_strategy}
\alias{complete_intake_strategy}
\title{Complete intake according to platform's strategy}
\usage{
complete_intake_strategy(data, intake, ...)
}
\arguments{
\item{data}{data.frame() survey data with all participants (typically weekly), see \code{\link{complete_intake}}}

\item{intake}{data.frame() intake data loaded on same period as data}

\item{...}{other parameters to pass to \code{\link{complete_intake}}}
}
\description{
The completion strategy is defined as a list of parameters under the `complete.intake` name in the platform options (see \code{\link{platform_options}})
This list is :
\describe{
 \item{max.year}{integer, maximum number of season to load to get a intake from the previous season}
}
}
