% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{create_binary_mask}
\alias{create_binary_mask}
\title{Create a binary mask for each value}
\usage{
create_binary_mask(values)
}
\arguments{
\item{values}{character}
}
\value{
vector of integer value as binary mask for each input values, named by input value
}
\description{
Create a binary mask for each value
}
