% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episode_build}
\alias{episode_build}
\title{Build Episodes data from survey data}
\usage{
episode_build(
  design,
  intake = NULL,
  weekly = NULL,
  syndrome.column,
  env = NULL
)
}
\arguments{
\item{design}{a design structure holding computation parameters, \code{\link{episode_design}}}

\item{intake}{intake data}

\item{weekly}{weekly data}

\item{syndrome.column}{name of the boolean column indicating if each weekly fit the syndrome definition or not (TRUE is yes)}

\item{env}{environment}
}
\value{
environment containing intake, weekly, participants
}
\description{
Build Episodes data from survey data
}
\examples{
\dontrun{
 episode_build(design, weekly, intake)
}

}
\seealso{
Other episodes: 
\code{\link{episode_compute}()},
\code{\link{episode_design}()},
\code{\link{episode_fusion}()},
\code{\link{episode_fusion.simple_strategy}()},
\code{\link{episode_fusion.worst_strategy}()},
\code{\link{episode_fusion_strategy}()},
\code{\link{episode_prepare_data}()},
\code{\link{episode_select_participants}()},
\code{\link{episode_strategy}()}
}
\concept{episodes}
