% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episode_compute}
\alias{episode_compute}
\title{Compute episodes in weekly data}
\usage{
episode_compute(.env, syndrome.column, design, .progress = TRUE, local = FALSE)
}
\arguments{
\item{.env}{environment result of \code{\link{episode_prepare_data}}}

\item{syndrome.column}{name of the column containing episode data}

\item{design}{study params, structure returned by \code{\link{episode_design}}}

\item{.progress}{optional progress bar structure, if not NULL must use dplyr::progress_estimated api}

\item{local}{if TRUE do not modify .env, only returns episodes}
}
\value{
data.frame (episodes)
}
\description{
This function use environment returned by \code{\link{episode_prepare_data}}
}
\details{
It is an intermediate function, in a simple setting \code{\link{episode_build}} can be more convenient to use
}
\seealso{
Other episodes: 
\code{\link{episode_build}()},
\code{\link{episode_design}()},
\code{\link{episode_fusion.simple_strategy}()},
\code{\link{episode_fusion.worst_strategy}()},
\code{\link{episode_fusion_strategy}()},
\code{\link{episode_fusion}()},
\code{\link{episode_prepare_data}()},
\code{\link{episode_select_participants}()},
\code{\link{episode_strategy}()}
}
\concept{episodes}
