% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_fusion.R
\name{episode_fusion.worst_strategy}
\alias{episode_fusion.worst_strategy}
\title{Apply 'worst' fusion strategy}
\usage{
\method{episode_fusion}{worst_strategy}(strategy, weekly, episode.column, ...)
}
\arguments{
\item{strategy}{strategy definition structure, created by \code{\link{episode_strategy}}}

\item{weekly}{weekly data.frame for one episode}

\item{episode.column}{episode column}

\item{...}{other arguments (not used)}
}
\value{
data.frame() one row by person_id, episode, with selected value for each variable registred in strategy
}
\description{
This strategy will consider
}
\seealso{
\code{\link{episode_fusion}}

Other episodes: 
\code{\link{episode_build}()},
\code{\link{episode_compute}()},
\code{\link{episode_design}()},
\code{\link{episode_fusion}()},
\code{\link{episode_fusion.simple_strategy}()},
\code{\link{episode_fusion_strategy}()},
\code{\link{episode_prepare_data}()},
\code{\link{episode_select_participants}()},
\code{\link{episode_strategy}()}
}
\concept{episodes}
