% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episode_prepare_data}
\alias{episode_prepare_data}
\title{Prepare data and select participants using given rules}
\usage{
episode_prepare_data(
  design,
  intake = NULL,
  weekly = NULL,
  env = NULL,
  keep.onset = FALSE
)
}
\arguments{
\item{design}{episode design data structure (result of \code{\link{episode_design}})}

\item{intake}{data.frame intake data}

\item{weekly}{data.frame weekly data}

\item{env}{environment with intake & weekly data (alternative to provide them directly in existing env)}

\item{keep.onset}{keep onset column if already exists}
}
\value{
environment environment containing selected data (weekly, intake, participants, selections)
}
\description{
`episode_prepare_data` makes some checks in data, and apply selection rules by callng \code{\link{episode_select_participants}}.
}
\details{
To compute episode in a simple setting \code{link{episode_build}} would be more convinient to use (it will call this function)


The environment contains several data.frame()
\describe{
 \item{participants}{one row for each participants and rule results (TRUE/FALSE) for each participants, column "keep" indicate final state of participant}
 \item{selections}{count of selected participants at each step (after applying each rule)}
 \item{intake}{the intakes for selected participants}
 \item{weekly}{the weekly for selected participants}
}
}
\seealso{
Other episodes: 
\code{\link{episode_build}()},
\code{\link{episode_compute}()},
\code{\link{episode_design}()},
\code{\link{episode_fusion}()},
\code{\link{episode_fusion.simple_strategy}()},
\code{\link{episode_fusion.worst_strategy}()},
\code{\link{episode_fusion_strategy}()},
\code{\link{episode_select_participants}()},
\code{\link{episode_strategy}()}
}
\concept{episodes}
