% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_base_table}
\alias{geo_base_table}
\title{Get Base geographic tables
It should contains all levels, describe all upper levels from the base level}
\usage{
geo_base_table()
}
\description{
Get Base geographic tables
It should contains all levels, describe all upper levels from the base level
}
