% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_definition}
\alias{geo_definition}
\title{Get geographic structure description}
\usage{
geo_definition(name = NULL)
}
\arguments{
\item{name}{geographic level name}
}
\description{
Get geographic structure description
}
