% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_level}
\alias{geo_level}
\title{Create a geographic level object}
\usage{
geo_level(geo.name, hierarchy = NULL, ...)
}
\arguments{
\item{geo.name}{geographic level name (for example "nuts1", "nuts2")}

\item{hierarchy}{name of a level hierarchy}

\item{...}{extra information to store with level information}
}
\description{
The available levels depends on the platform configuration
}
