% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{geo_level_table}
\alias{geo_level_table}
\title{Get geographic table definition for a given geographic level}
\usage{
geo_level_table(level)
}
\arguments{
\item{level}{geographic level name}
}
\description{
Get geographic table definition for a given geographic level
}
\seealso{
Other geography: 
\code{\link{geo_base_level}()},
\code{\link{geo_base_table}()},
\code{\link{geo_definition}()},
\code{\link{geo_hierarchy}()},
\code{\link{geo_is_type}()},
\code{\link{geo_level_nav}()},
\code{\link{geo_level}()},
\code{\link{get_geo_join}()},
\code{\link{load_geo_zone}()},
\code{\link{load_population_age}()},
\code{\link{load_population}()}
}
\concept{geography}
