% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{get_geo_join}
\alias{get_geo_join}
\title{Get information needed to make a join with geo tables
from base geographic level to any other level}
\usage{
get_geo_join()
}
\value{
list()

\describe{
 \item{table}{name of the table associating all levels}
 \item{base.column}{name of the column for the lowest level}
 \item{join.country}{TRUE if the join should include the "country" column (european data or cross country site)}
}
}
\description{
Get information needed to make a join with geo tables
from base geographic level to any other level
}
