% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_historical_tables}
\alias{get_historical_tables}
\title{Get Historical tables}
\usage{
get_historical_tables()
}
\description{
Historical tables is a list, with season year (first year of the season) as name and season definition
}
\seealso{
Season in \code{\link{concepts}}

season.def

Other seasons: 
\code{\link{get_historical_seasons}()},
\code{\link{get_season_dates}()},
\code{\link{season_definition}()}
}
\concept{seasons}
