% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.R
\name{get_option}
\alias{get_option}
\title{Get package option from name
From 'ifn' options() entry}
\usage{
get_option(name = NULL)
}
\description{
Get package option from name
From 'ifn' options() entry
}
\seealso{
output_options
}
