% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{graph.open}
\alias{graph.open}
\title{Open a graph using standard library}
\usage{
graph.open(file, width = NA, height = NA, pitch = 12, type = "png",
  ...)
}
\arguments{
\item{file}{filename (without extension)}

\item{width}{width (pixel)}

\item{height}{height in px}

\item{pitch}{pointsize value}

\item{type}{type of output to use}

\item{...}{extra parameted}
}
\description{
Open a graph using standard library
}
