% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.r
\name{i18n_set}
\alias{i18n_set}
\title{Register new translations from a list}
\usage{
i18n_set(...)
}
\arguments{
\item{...}{named value for text to be translated}
}
\description{
Register new translations from a list
}
