% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_rs_2014_daily.R
\docType{data}
\name{IncidenceDailyRS2014}
\alias{IncidenceDailyRS2014}
\title{Incidence estimator using rs2014 method implemented using daily agregatation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IncidenceDailyRS2014
}
\description{
Incidence estimator using rs2014 method implemented using daily agregatation
}
\details{
Input:
Input data are expected to have some columns & some fixes. They are loaded and prepared using \code{\link{load_results_for_incidence}}

Parameters:

\describe{
 \item{ignore.first.delay}{Number of days form his first survey befotre to include a participant (not compatible with ignore.first)}
 \item{ignore.first.only.new}{Ignore first rule only for new participant (uses intake$first.season column}
 \item{exclude.same}{should exclude syndrom if same episode is on (partcipant still active)}
 \item{exclude.same.delay}{max delay to exclude syndrom if same (not excluded if over this delay)}
 \item{active.max.freq}{max delay between 2 surveys (in weeks)}
 \item{active.day.before}{Number of week before each computed week to count active participants}
 \item{active.day.after}{Number of week after each computed week to count active participants}
 \item{active.min.surveys}{min number of surveys for each participant (not active if less)}
 \item{active.use.mean}{Use mean number of active participant over the windows [active.day.before, active.day.after]}
}
}
\section{Fields}{

\describe{
\item{\code{weekly}}{weekly data (loaded using \code{\link{load_results_for_incidence}})}

\item{\code{intake}}{intake data (loaded using \code{\link{load_results_for_incidence}})}

\item{\code{participants}}{data.frame() with all available participants and commputed criterias used during computation}

\item{\code{params}}{parameters for computation}

\item{\code{syndroms}}{character vector of column names containing syndroms classification for each weekly}

\item{\code{verbose}}{logical show verbose message}

\item{\code{design}}{design stratification from \code{\link{design_incidence}}}

\item{\code{output}}{vector of character, see \code{\link{IncidenceRS2014}}}
}}

\keyword{datasets}
