% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_rs_2014.R
\docType{class}
\name{IncidenceRS2014}
\alias{IncidenceRS2014}
\title{Compute Incidence using rs2014 method
Active participants are selected using a set of rules}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IncidenceRS2014
}
\description{
Compute Incidence using rs2014 method
Active participants are selected using a set of rules
}
\details{
Parameters:
\describe{
 \item{ignore.first.delay}{Number of days form his first survey befotre to include a participant (not compatible with ignore.first)}
 \item{ignore.first.only.new}{Ignore first rule only for new participant (uses intake$first.season column @see load_results_for_incidence)}
 \item{exclude.same}{should exclude syndrom if same episode is on (partcipant still active)}
 \item{exclude.same.delay}{max delay to exclude syndrom if same (not excluded if over this delay)}
 \item{active.max.freq}{max delay between 2 surveys (in weeks)}
 \item{active.week.before}{Number of week before each computed week to count active participants}
 \item{active.week.after}{Number of week after each computed week to count active participants}
 \item{active.min.surveys}{min number of surveys for each participant (not active if less)}
}

Input:
Input data are expected to have some columns & some fixes. They are loaded and prepared using \code{\link{load_results_for_incidence}}

Output:
Incidence estimator can output several kind of datasets
\describe{
 \item{inc}{inc=incidence at (national level)}
 \item{zlow}{"lower geographic level (z)", if estimator is stratified by geographic level}
 \item{age}{"age" age-specific incidence, age categories should be in "age.cat" column in intake data}
 \item{count}{raw count of syndroms used for each syndrom by week, before incidence is computed}
}

Apply the full algorithm for rs2014 incidence computation
}
\section{Fields}{

\describe{
\item{\code{weekly}}{weekly data}

\item{\code{intake}}{intake data}

\item{\code{participants}}{data.frame() with all available participants and commputed criterias used during computation}

\item{\code{params}}{parameters for computation}

\item{\code{syndroms}}{character vector of column names containing syndroms classification for each weekly}

\item{\code{profiler}}{profiler}

\item{\code{design}}{design stratification from \code{\link{design_incidence}}}

\item{\code{output}}{vector of character (see Details outupus)}
}}

\keyword{datasets}
