% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{init.path}
\alias{init.path}
\title{Define an ouput path for files, usable by my.path()
by default, the path is added to the global output path (defined in OUT_PATH), unless the parameter full.path is TRUE}
\usage{
init.path(p, full.path = F)
}
\arguments{
\item{p}{chr path}

\item{full.path}{logical if TRUE p is considered as an absolute path}
}
\description{
Define an ouput path for files, usable by my.path()
by default, the path is added to the global output path (defined in OUT_PATH), unless the parameter full.path is TRUE
}
\seealso{
Other path functions: \code{\link{my.path}}
}
