% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{load_geo_zone}
\alias{load_geo_zone}
\title{Load areas defined at a given geographic level}
\usage{
load_geo_zone(geo, type = "metro", columns = c())
}
\arguments{
\item{geo}{geographic level code (ex "nuts1")}

\item{type}{type of geographic level (if needed)}

\item{columns}{list of extra columns to fetch (if the geographic table has extra columns...)}
}
\description{
Load areas defined at a given geographic level
}
\seealso{
Other geography: 
\code{\link{geo_base_level}()},
\code{\link{geo_base_table}()},
\code{\link{geo_definition}()},
\code{\link{geo_hierarchy}()},
\code{\link{geo_is_type}()},
\code{\link{geo_level}()},
\code{\link{geo_level_nav}()},
\code{\link{geo_level_table}()},
\code{\link{get_geo_join}()},
\code{\link{load_population}()},
\code{\link{load_population_age}()}
}
\concept{geography}
