% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{load_population}
\alias{load_population}
\alias{load_population.db}
\alias{load_population.age}
\title{Load global population for a geographic level and a census year}
\usage{
load_population(geo, year, country = NULL)

load_population.db(geo, year, country)

load_population.age(geo, year, country)
}
\arguments{
\item{geo}{geographic level}

\item{year}{year of the population (estimation of the population for the year)}

\item{country}{country (only on country enabled platform)}
}
\description{
Population table should follow the structure described in geography vignette
}
\seealso{
Other geography: 
\code{\link{geo_base_level}()},
\code{\link{geo_base_table}()},
\code{\link{geo_definition}()},
\code{\link{geo_hierarchy}()},
\code{\link{geo_is_type}()},
\code{\link{geo_level_nav}()},
\code{\link{geo_level_table}()},
\code{\link{geo_level}()},
\code{\link{get_geo_join}()},
\code{\link{load_geo_zone}()},
\code{\link{load_population_age}()}

Other geography: 
\code{\link{geo_base_level}()},
\code{\link{geo_base_table}()},
\code{\link{geo_definition}()},
\code{\link{geo_hierarchy}()},
\code{\link{geo_is_type}()},
\code{\link{geo_level_nav}()},
\code{\link{geo_level_table}()},
\code{\link{geo_level}()},
\code{\link{get_geo_join}()},
\code{\link{load_geo_zone}()},
\code{\link{load_population_age}()}
}
\concept{geography}
