% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{load_population}
\alias{load_population}
\title{Load global population for a geographic level and a census year}
\usage{
load_population(geo, year)
}
\arguments{
\item{geo}{geographic level}

\item{year}{year of the population (estimation of the population for the year)}
}
\description{
Population table should follow the structure described in geography vignette
}
