% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{load_population_age.impl}
\alias{load_population_age.impl}
\title{Load Population by age}
\usage{
load_population_age.impl(
  loader_type,
  geo,
  year,
  age.breaks = NULL,
  type = NULL,
  country = NULL,
  ...
)
}
\arguments{
\item{loader_type}{type of loader to use}

\item{geo}{geographic level, if null, return by age only population}

\item{year}{year of the population}

\item{age.breaks}{used to create age categories (from 5-year categories)}

\item{...}{extra parameter}
}
\value{
data.frame('all','male','female','age.cat') for the given year
}
\description{
Load popuplation data by 5-year (or multiple) age group & sex
}
\details{
3 loaders are availables :
\describe{
 \item{db}{Database loader}
 \item{file}{File for all the platform}
 \item{country_file}{File by country, this require the platform variable "country" to be defined to the country to use}
}
}
