% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{load_population_age}
\alias{load_population_age}
\title{Load popuplation by age-group & sex}
\usage{
load_population_age(geo, year, age.breaks = NULL, version = NULL, ...)
}
\arguments{
\item{geo}{geographic level, if null, return by age only population}

\item{year}{year of the population}

\item{age.breaks}{used to create age categories (from 5-year categories)}

\item{version}{version of the file to use as data source}

\item{...}{extra parameter}
}
\value{
data.frame('all','male','female','age.cat') for the given year
}
\description{
Load popuplation data by 5-year (or multiple) age group & sex
}
\seealso{
Other geography: 
\code{\link{geo_base_level}()},
\code{\link{geo_base_table}()},
\code{\link{geo_definition}()},
\code{\link{geo_hierarchy}()},
\code{\link{geo_is_type}()},
\code{\link{geo_level}()},
\code{\link{geo_level_nav}()},
\code{\link{geo_level_table}()},
\code{\link{get_geo_join}()},
\code{\link{load_geo_zone}()},
\code{\link{load_population}()}
}
\concept{geography}
