% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.r
\name{max_year_available}
\alias{max_year_available}
\title{Allow the population year to be the last available}
\usage{
max_year_available(year)
}
\arguments{
\item{year}{requested year}
}
\description{
This function wrap the requested year number adding attribute to inform the population
loader that it can provide population of previous years if the requested year is not available
This allow to use an older population if the population data source is not up to date yet.
}
