% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{msg}
\alias{msg}
\title{Wrapper of message() for output
You car redirect by using sink() function}
\usage{
msg(..., ln = TRUE, level = 0)
}
\arguments{
\item{...}{message (multiple will be concatenated)}

\item{ln}{add new line}

\item{level}{level if hierarchical levels handled (ignored if not)
Deprecated}
}
\description{
Wrapper of message() for output
You car redirect by using sink() function
}
