% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\docType{data}
\name{NO_SYMPTOM}
\alias{NO_SYMPTOM}
\title{No symptome variable name}
\format{
An object of class \code{character} of length 1.
}
\usage{
NO_SYMPTOM
}
\description{
To be used as a constant in the analysis
}
\keyword{datasets}
