% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_design.R
\name{output_incidence}
\alias{output_incidence}
\title{Define structure containing incidence results output and some metadata}
\usage{
output_incidence(types, conf.int = T, adjust = T, ...)
}
\arguments{
\item{types}{types of measures computed}

\item{conf.int}{confidence interval computed}

\item{adjust}{adjusted measures}

\item{...}{data sets}
}
\description{
This function which allow to pass some extra parameters to the output function \code{\link{calc_adjusted_incidence}}
}
