% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{override}
\alias{override}
\title{Allow override of mapping or recoding from template}
\usage{
override(data)
}
\arguments{
\item{data}{value to flag as overrided}
}
\description{
Allow override of mapping or recoding from template
}
