% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform_define_survey}
\alias{platform_define_survey}
\title{Describe a survey of a platform and register it}
\usage{
platform_define_survey(
  name,
  survey_id = NULL,
  table = NULL,
  mapping = list(),
  labels = NULL,
  codes = NULL,
  recodes = list(),
  single.table = FALSE,
  geo.column = NULL,
  template = NULL,
  ...
)
}
\arguments{
\item{name}{unique name of the survey}

\item{survey_id}{id of the survey in the database}

\item{table}{name of the default table storing response of the survey}

\item{mapping}{Mapping from DB names (in value) and R names for each survey variables.}

\item{labels}{list of labels}

\item{codes}{list of codes *deprecated*}

\item{recodes}{list of recoding}

\item{single.table}{boolean, TRUE if survey uses single table model}

\item{geo.column}{name of the variable (R names)}

\item{template}{name of the template to override}

\item{...}{extra parameters to set for the survey}
}
\value{
invisibly return the survey_definition object
}
\description{
Describe configuration of a survey and the mapping from DB names &
coding to R names & factor levels.
A call to this function is needed to register a survey in the package before to use other survey functions.
Usually it is done in the platform definition file, which is loaded by \code{\link{load_platform}()} function
}
\section{Mapping}{


The mapping allows to automatically transform db names to R names (more meaningfull and error-proof) each times data are loaded using `survey_load_results()`
}

\section{Recoding}{


for each variable, it is possible to give a list(label1=code1, label2=code2) where `code1`, `code2` will be recoded to the corresponding label
The idea is to give a human meaningful label to each numeric code used to store response value for questions based on an option list.

It is a good pratice to use variable style names (character + dot), without space and in english. Those labels could be translated
later to another language in the analysis output (graphs, tables & so on). This aimed to produced shareable and language agnostic
programs.
For a given questions, all options should also start with the same prefix to allow quick selection of theses labels based on patterns

For example:
 main.activity = c('activity.fulltime'='0','activity.partial'='1','activity.self'='2', 'activity.student'='3','activity.home'='4','activity.unemployed'='5','activity.sick'='6','activity.retired'='7', 'activity.other'=8),

Will describe recoding for the question named 'main.activity' (using attributed name, never the DB name).
}

\section{Labels}{


labels are named list of labels (a label is just a character string). It is used to manipulate list of names for various purposes : list for variables
for multi-valued questions (from "checkboxes"), a label provide a "name" to identify the list of all variables corresponding to one question.
It is also used to get the list of recoded labels for a qualitative question.

One of common labels is "symptoms" in weekly survey definition, providing the list of the variables names for symptoms question.
}

\seealso{
\code{\link{survey_definition}()}

\code{\link{survey_recode}}

Other platform: 
\code{\link{create_survey_definition}()},
\code{\link{load_platform}()},
\code{\link{platform_import}()}
}
\concept{platform}
