% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform_geographic_levels}
\alias{platform_geographic_levels}
\title{Create structure defining an geographic table system}
\usage{
platform_geographic_levels(
  levels,
  level.base = NULL,
  table = "geo_levels",
  columns = NULL,
  hierarchies = NULL,
  default.hierarchy = "default",
  country = FALSE,
  define = TRUE
)
}
\arguments{
\item{levels}{list of levels with level name in names and column name in value, if it not a named vector, the use value as level}

\item{level.base}{name of the base level (if null, first level is used)}

\item{table}{name of the table containing all the levels (it can be the lowest level table as it should contains reference to all upper levels)}

\item{columns}{list of column name for each level in the form list([name]=[column_name])}

\item{hierarchies}{list of hierarchies list( hierarchy1=c(levels order), hierarchy2=(levels order), ...)}

\item{default.hierarchy}{name of the default hierarchy to be used to get upper or lower column}

\item{country}{handle country column}

\item{define}{if TRUE set the geo.levels during the call, set to FALSE if you dont want to change configuration}
}
\description{
Create structure defining an geographic table system
}
