% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform_options}
\alias{platform_options}
\title{Define  platform options}
\usage{
platform_options(...)
}
\arguments{
\item{...}{list of options to set

\describe{
  \item{first.season.censored}{left censor first season participants for some countries}
  \item{health.status}{structure of the health.status table in case of single table model for weekly}
  \item{debug.query}{debug SQL queries}
  \item{use.country}{If the platform handles country level in the database tables}
  \item{population.age.loader}{Loader type for age-group population, either "db","file" or a function with \code{\link{load_population_age}()} interface}
  \item{population.loader}{Loader type for overall population, either "db","file" or a function with \code{\link{load_population}()} interface}
  \item{language}{Language code to use by default when loading i18n}
  \item{complete.intake}{list defining intake completion strategy see \code{\link{complete_intake_strategy}()} }
}}
}
\description{
Define  platform options
}
\seealso{
Other platfom: 
\code{\link{platform_env}()},
\code{\link{platform_geographic_levels}()},
\code{\link{platform_geographic_tables}()},
\code{\link{platform_season_history}()}
}
\concept{platfom}
