% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{platform_path}
\alias{platform_path}
\title{Path to platform definition files}
\usage{
platform_path(file = "")
}
\description{
Path to platform definition files
}
\seealso{
Other path-functions: 
\code{\link{add_path_prefix}()},
\code{\link{get_current_paths}()},
\code{\link{init.path}()},
\code{\link{my.path}()},
\code{\link{print.paths_definition}()},
\code{\link{share.cache.path}()},
\code{\link{share.data.path}()},
\code{\link{share.lib.path}()},
\code{\link{share_path}()}
}
\concept{path-functions}
