% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{platform_path}
\alias{platform_path}
\title{Path to platform definition files}
\usage{
platform_path(file = "")
}
\arguments{
\item{file}{character file name to get in the 'platform' directory}
}
\value{
character path of the file in platform files location.
}
\description{
Platform files are located in the path defined by the option "platform.path" (see \code{\link{share.option}})
Usually in share/platform in the root of the workspace (see \code{\link{concepts}}).
}
\seealso{
Other path-functions: 
\code{\link{add_path_prefix}()},
\code{\link{get_current_paths}()},
\code{\link{init.path}()},
\code{\link{my.path}()},
\code{\link{print.paths_definition}()},
\code{\link{share.cache.path}()},
\code{\link{share.data.path}()},
\code{\link{share.lib.path}()},
\code{\link{share_path}()}
}
\concept{path-functions}
