% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{print.survey_error}
\alias{print.survey_error}
\title{Print checks results}
\usage{
\method{print}{survey_error}(x, ...)
}
\arguments{
\item{x}{list() errors generated by \code{\link{create_survey_definition}}}

\item{...}{extra parameters (for print interface compatibility)}
}
\description{
Print checks results
}
