% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{prop_by_weighted}
\alias{prop_by_weighted}
\title{Compute weighted proportion by group for a logical variable}
\usage{
prop_by_weighted(design, var, by, ci.method = "beta")
}
\arguments{
\item{design}{survey design [survey::svydesign()], see [design_weight()]}

\item{var}{variable on which to compute frequency, must be a logical variable}

\item{by}{grouping variable name to compute total on}

\item{ci.method}{method to use to compute adjusted proportion confidence interval}
}
\value{
data.frame()
   with columns
   \describe{
    \item{by...}{group name of this column is the passed value to the `by` parameter}
    \item{n_adj}{Count of true values in each group}
    \item{n_weight}{Weight sum of the true value in the group}
    \item{total_weight}{Total weight in the group}
    \item{total_adj}{Total count of non missing value in the group}
    \item{prop_adj}{Adjusted proportion on strata}
    \item{prop_adj_low}{Adjusted proportion lower bound of CI95\%}
    \item{prop_adj_up}{Adjusted proportion upper bound of CI95\%}
    \item{variable}{Variable name}
   }
}
\description{
Compute weighted proportion by group for a logical variable
}
