% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_utils.R
\name{recode_alias}
\alias{recode_alias}
\title{create a recoding alias}
\usage{
recode_alias(name)
}
\arguments{
\item{name}{name of the variable to use the recoding from}
}
\description{
The recoding will be proceeded using the name provided. This avoid to copy paste the recode mapping
by reusing the one defined in the same survey.
}
