% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{relative_seasons}
\alias{relative_seasons}
\title{Get relative season numbers to a reference season number}
\usage{
relative_seasons(season, index = NULL, .all = NULL)
}
\arguments{
\item{season}{int reference season}

\item{index}{int vector of relative index to the reference season, if NA or NULL get all previous}

\item{.all}{int[] only select on this seasons, if not NULL}
}
\value{
vector of season numbers matching the index relative to season reference
}
\description{
Get relative season numbers to a reference season number
}
\keyword{internal}
