% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{share.cache.path}
\alias{share.cache.path}
\title{return the path of a file in cache files}
\usage{
share.cache.path(file = "")
}
\arguments{
\item{file}{file name to load}
}
\description{
return the path of a file in cache files
}
\seealso{
Other path-functions: 
\code{\link{platform_path}()},
\code{\link{share.data.path}()},
\code{\link{share.lib.path}()},
\code{\link{share_path}()}
}
\concept{path-functions}
