% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{share_path}
\alias{share_path}
\alias{share.path}
\title{path of a file in share/ directory}
\usage{
share_path(file = "")

share.path(file = "")
}
\arguments{
\item{file}{file name to load}
}
\description{
path of a file in share/ directory
}
\seealso{
Other path-functions: 
\code{\link{add_path_prefix}()},
\code{\link{create_path}()},
\code{\link{get_current_paths}()},
\code{\link{init.path}()},
\code{\link{my.path}()},
\code{\link{platform_path}()},
\code{\link{print.paths_definition}()},
\code{\link{share.cache.path}()},
\code{\link{share.data.path}()},
\code{\link{share.lib.path}()}
}
\concept{path-functions}
