% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_aliases}
\alias{survey_aliases}
\title{Convert column name (InfluenzaNet column names in the db) from (revert=F) and to variable aliases (revert=T)}
\usage{
survey_aliases(cols, def, revert = F)
}
\arguments{
\item{cols}{names of columns to rename}

\item{def}{list of aliases}

\item{revert}{if TRUE, aliases to Db names, if false (default) DB names to aliases}
}
\description{
Description of mapping are defined in the survey  by \code{\link{platform_define_survey}} in the platform file
}
