% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_definition}
\alias{survey_definition}
\title{Get survey definition
Survey definition is a data structure of entries defining variable mapping (db column name to variable name, recoding, variable sets)
They are defined by \code{\link{platform_define_survey}} usually in the platform file}
\usage{
survey_definition(survey)
}
\arguments{
\item{survey}{survey name}
}
\value{
list parameters describing a survey
}
\description{
Get survey definition
Survey definition is a data structure of entries defining variable mapping (db column name to variable name, recoding, variable sets)
They are defined by \code{\link{platform_define_survey}} usually in the platform file
}
\seealso{
Other survey: 
\code{\link{survey_aliases}()},
\code{\link{survey_labels}()},
\code{\link{survey_recode}()},
\code{\link{survey_recodings}()},
\code{\link{survey_variable_recoding}()},
\code{\link{survey_variables_like}()}
}
\concept{survey}
