% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{survey_load_participants}
\alias{survey_load_participants}
\title{Load participants data}
\usage{
survey_load_participants(active.account = NULL, ids = NULL)
}
\arguments{
\item{active.account}{logical only active user account if TRUE}

\item{ids}{list of survey_user ids}
}
\description{
Load participants data
}
\seealso{
Other survey-load: 
\code{\link{survey_load_participations}()},
\code{\link{survey_load_results}()},
\code{\link{survey_load_results_historic}()},
\code{\link{survey_participant_previous_season}()},
\code{\link{survey_participant_season}()}
}
\concept{survey-load}
