% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{survey_load_participations}
\alias{survey_load_participations}
\title{Load last participation date for intake & weekly survey for each participants}
\usage{
survey_load_participations(ids, years = NULL)
}
\arguments{
\item{ids}{list of participants}

\item{years}{list of season to scan, all if NULL}
}
\description{
Load last participation date for intake & weekly survey for each participants
}
\seealso{
Other survey-load: 
\code{\link{survey_load_participants}()},
\code{\link{survey_load_results}()},
\code{\link{survey_load_results_historic}()},
\code{\link{survey_participant_previous_season}()},
\code{\link{survey_participant_season}()}
}
\concept{survey-load}
