% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_load_questions}
\alias{survey_load_questions}
\title{load questions of a survey from the database}
\usage{
survey_load_questions(survey, translation_id = NULL)
}
\arguments{
\item{survey}{survey name (@see epiwork.tables)}

\item{translation_id}{id of the translation to use}
}
\description{
load questions of a survey from the database
}
\seealso{
Other survey definition functions: \code{\link{survey_load_options}}
}
\concept{survey definition functions}
