% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_loader.R
\name{survey_load_translation}
\alias{survey_load_translation}
\title{load a translation for a survey for a given language}
\usage{
survey_load_translation(survey, language = survey_default_language())
}
\arguments{
\item{survey}{survey name}

\item{language}{code (as used in the db)}
}
\description{
load a translation for a survey for a given language
}
\seealso{
Other survey-load: 
\code{\link{survey_load_all}()},
\code{\link{survey_load_options}()},
\code{\link{survey_load_participants}()},
\code{\link{survey_load_participations}()},
\code{\link{survey_load_questions}()},
\code{\link{survey_load_results}()},
\code{\link{survey_load_results_historic}()},
\code{\link{survey_options_for}()},
\code{\link{survey_participant_previous_season}()},
\code{\link{survey_participant_season}()},
\code{\link{survey_question_id}()}
}
\concept{survey-load}
