% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_options_for}
\alias{survey_options_for}
\title{Get the options for a survey's varname
This is not the "options" defined}
\usage{
survey_options_for(survey, varname)
}
\arguments{
\item{survey}{survey name}

\item{varname}{variable name}
}
\description{
Get the options for a survey's varname
This is not the "options" defined
}
