% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_participant_previous_season}
\alias{survey_participant_previous_season}
\title{List of participants registred in weekly at least once in previous season (regarding given [season])}
\usage{
survey_participant_previous_season(season, ids = NULL,
  use.season.dates = F, seasons = NULL)
}
\arguments{
\item{season}{season year}

\item{ids}{list of participants to keep}

\item{use.season.dates}{if TRUE restrist weekly scan to the official date of each season (@see historical.tables)}

\item{seasons}{list of seasons to scan, relatively to the given [season], for ex -1 = only previous}
}
\description{
List of participants registred in weekly at least once in previous season (regarding given [season])
}
