% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{survey_participant_previous_season}
\alias{survey_participant_previous_season}
\alias{survey_participant_previous_season.multiple_table}
\alias{survey_participant_previous_season.single_table}
\title{List of participants registered in weekly at least once in previous season (regarding given [season])}
\usage{
survey_participant_previous_season(
  season,
  ids = NULL,
  use.season.dates = F,
  from = NULL,
  country = NULL,
  verbose = FALSE
)

survey_participant_previous_season.multiple_table(
  season,
  ids = NULL,
  use.season.dates = F,
  from = NULL,
  country = NULL,
  verbose = TRUE
)

survey_participant_previous_season.single_table(
  season,
  ids = NULL,
  from = NULL,
  country = NULL,
  verbose = TRUE
)
}
\arguments{
\item{season}{season year}

\item{ids}{list of participants to keep}

\item{use.season.dates}{if TRUE restrict weekly scan to the official date of each season \code{\link{get_historical_tables}}}

\item{from}{integer relative index of the oldest season to scan to (index relative to season, e.g. -1 = previous from given [season])}

\item{country}{country to use (only on european platform)}

\item{verbose}{show verbose messages}
}
\description{
List of participants registered in weekly at least once in previous season (regarding given [season])

Implementation for multiple table data model
we have to check season by season

Implementation for multiple table data model
we have to check season by season
}
\seealso{
Other survey-load: 
\code{\link{survey_load_participants}()},
\code{\link{survey_load_participations}()},
\code{\link{survey_load_results}()},
\code{\link{survey_load_results_historic}()},
\code{\link{survey_participant_season}()}
}
\concept{survey-load}
