% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_load_data.R
\name{survey_participant_season}
\alias{survey_participant_season}
\title{Get list of participant id (person_id = survey_user_id)
List of participants registered in weekly at least once for a given season}
\usage{
survey_participant_season(
  season,
  use.season.dates = FALSE,
  use.min = TRUE,
  country = NULL
)
}
\arguments{
\item{season}{season number to get, if several seasons are given use min and max of seasons dates}

\item{use.season.dates}{restrict to season's starting & ending dates. forced if single table model for weekly}

\item{use.min}{us minimal date of the given season, if FALSE use all before the end of the season (only for single table model)}

\item{country}{country to restrict participant}
}
\description{
Get list of participant id (person_id = survey_user_id)
List of participants registered in weekly at least once for a given season
}
\seealso{
Other survey-load: 
\code{\link{survey_load_all}()},
\code{\link{survey_load_options}()},
\code{\link{survey_load_participants}()},
\code{\link{survey_load_participations}()},
\code{\link{survey_load_questions}()},
\code{\link{survey_load_results_historic}()},
\code{\link{survey_load_results}()},
\code{\link{survey_load_translation}()},
\code{\link{survey_options_for}()},
\code{\link{survey_participant_previous_season}()},
\code{\link{survey_question_id}()}
}
\concept{survey-load}
