% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_loader.R
\name{survey_question_id}
\alias{survey_question_id}
\title{get the question_id (id in the question table in the influenzanet db) of a question}
\usage{
survey_question_id(survey, varname, language = survey_default_language())
}
\arguments{
\item{survey}{name of the survey}

\item{varname}{variable name of the question (db's column name or R alias of the column)}

\item{language}{language translation to use}
}
\value{
int
}
\description{
get the question_id (id in the question table in the influenzanet db) of a question
}
