% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_questions_like}
\alias{survey_questions_like}
\title{Returns list of question aliases matching the given pattern (glob style)}
\usage{
survey_questions_like(survey, pattern)
}
\arguments{
\item{survey}{survey name}

\item{pattern}{glob style pattern (ex "visit.*")}
}
\description{
Returns list of question aliases matching the given pattern (glob style)
}
