% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_recode}
\alias{survey_recode}
\title{Recode data from the DB storage code to the R labels (or to its translaction into default language)}
\usage{
survey_recode(x, question, survey, translate = F)
}
\arguments{
\item{x}{values to recodes}

\item{question}{name of the question (alias here)}

\item{survey}{name of the survey}

\item{translate}{if TRUE, try to translate the labels (using i18n function @seealso i18n)}
}
\value{
vector of recoded value in factor
}
\description{
This function recode from database values to labels (more meaningful). The recoding mapping for each variable is
defined \code{\link{platform_define_survey}} usually in the platform file
}
