% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_recode}
\alias{survey_recode}
\title{Recode data from the DB storage code to the R labels (or to its translaction into default language)}
\usage{
survey_recode(x, variable, survey, translate = F, question = NULL)
}
\arguments{
\item{x}{values to recodes}

\item{variable}{name of the question variable name to use for recoding}

\item{survey}{name of the survey}

\item{translate}{if TRUE, try to translate the labels (using i18n function @seealso i18n)}

\item{question}{old parameter name for variable, for compatibility}
}
\value{
vector of recoded value in factor
}
\description{
This function recode from database values to labels (more meaningful). The recoding mapping for each variable is
defined \code{\link{platform_define_survey}} usually in the platform file
}
\seealso{
Other survey: 
\code{\link{survey_aliases}()},
\code{\link{survey_definition}()},
\code{\link{survey_labels}()},
\code{\link{survey_recodings}()},
\code{\link{survey_rename_columns}()},
\code{\link{survey_variable_recoding}()},
\code{\link{survey_variables_like}()}
}
\concept{survey}
