% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.r
\name{survey_recode_all}
\alias{survey_recode_all}
\title{Recode all known variables in survey data}
\usage{
survey_recode_all(data, survey, warn = FALSE)
}
\arguments{
\item{data}{data.frame with weekly data}

\item{survey}{survey name (as it is registered in \code{\link{platform_define_survey}})}
}
\value{
data with recoded variable (to factor)
}
\description{
Recode all known variables in survey data
}
