% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_recodings}
\alias{survey_recodings}
\title{Get all the recodings defined in a survey}
\usage{
survey_recodings(survey)
}
\arguments{
\item{survey}{character survey name}
}
\value{
list with question variable name in name, a mapping as value
}
\description{
Get all the recodings defined in a survey
}
