% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{survey_variable_recoding}
\alias{survey_variable_recoding}
\title{Get the recoding mapping of a variable in a survey}
\usage{
survey_variable_recoding(survey, name, must.exists = TRUE)
}
\arguments{
\item{survey}{character survey name}

\item{name}{variable name}

\item{must.exists}{if TRUE an error will be raised if name doesnt exists in survey definition}
}
\value{
vector with label as name, database value as value
}
\description{
Get the recoding mapping of a variable in a survey
}
\seealso{
Other survey: 
\code{\link{SurveyLabeller}},
\code{\link{survey_aliases}()},
\code{\link{survey_definition}()},
\code{\link{survey_labeller}()},
\code{\link{survey_labels}()},
\code{\link{survey_recode}()},
\code{\link{survey_recodings}()},
\code{\link{survey_rename_columns}()},
\code{\link{survey_variables_like}()}
}
\concept{survey}
