% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{symptoms.gastro.rs}
\alias{symptoms.gastro.rs}
\title{Compute gastroenteriritis definition according to French Sentinel's definition}
\usage{
symptoms.gastro.rs(r)
}
\description{
Compute gastroenteriritis definition according to French Sentinel's definition
}
