% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome_provider.R
\name{SyndromeProvider}
\alias{SyndromeProvider}
\title{Syndrome provider base class}
\description{
Syndrome provider base class

Syndrome provider base class
}
\details{
Define interface to build a SyndromeProvider

This class compute "syndrome" definition from weekly & intake data
A syndrome will be computed as a logical value attributed for each weekly survey, with value TRUE if the survey comply with the given syndrome definition

A syndrome provider can compute several syndromes, each one applying for a definition

compute() should return a data.frame() with one logical column for each computed syndrome definition (use names carefully to avoid names collisions, avoid using
influenzanet base syndrome names : ili, cold, allergy, gastro )
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-compute}{\code{SyndromeProvider$compute()}}
\item \href{#method-compute_age}{\code{SyndromeProvider$compute_age()}}
\item \href{#method-is_sudden}{\code{SyndromeProvider$is_sudden()}}
\item \href{#method-clone}{\code{SyndromeProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute"></a>}}
\subsection{Method \code{compute()}}{
Compute syndrome
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProvider$compute(weekly, intake)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weekly}}{weekly data}

\item{\code{intake}}{intake data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame() with id column from weekly and extra column for each syndrome
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_age"></a>}}
\subsection{Method \code{compute_age()}}{
Compute age for each weekly using intake
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProvider$compute_age(weekly, intake)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weekly}}{weekly data}

\item{\code{intake}}{intake data
Common function, compute sudden symptomes using both variables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_sudden"></a>}}
\subsection{Method \code{is_sudden()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProvider$is_sudden(r)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{weekly data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
