% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome_provider_country.R
\name{SyndromeProviderCountry}
\alias{SyndromeProviderCountry}
\title{Syndrome Provider}
\description{
Syndrome Provider

Syndrome Provider
}
\section{Super class}{
\code{\link[ifnBase:SyndromeProvider]{ifnBase::SyndromeProvider}} -> \code{SyndromeProviderCountry}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pain.age.limit}}{Age to take into account of pain. Under this age pain=TRUE (disabled if NA)}

\item{\code{country}}{Country code}

\item{\code{season}}{Season number \code{\link{concepts}}}

\item{\code{use.fever.level}}{Use fever level when definition request a fever level or only the fever=TRUE}

\item{\code{unknown.has.fever}}{if TRUE consider that unknown fever status is fever=TRUE}

\item{\code{min.syndrome.count}}{Minimum syndrom count to take into account when definition doesnt say it}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SyndromeProviderCountry$new()}}
\item \href{#method-compute}{\code{SyndromeProviderCountry$compute()}}
\item \href{#method-clone}{\code{SyndromeProviderCountry$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ifnBase" data-topic="SyndromeProvider" data-id="compute_age">}\href{../../ifnBase/html/SyndromeProvider.html#method-compute_age}{\code{ifnBase::SyndromeProvider$compute_age()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ifnBase" data-topic="SyndromeProvider" data-id="is_sudden">}\href{../../ifnBase/html/SyndromeProvider.html#method-is_sudden}{\code{ifnBase::SyndromeProvider$is_sudden()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ifnBase" data-topic="SyndromeProvider" data-id="syndromes">}\href{../../ifnBase/html/SyndromeProvider.html#method-syndromes}{\code{ifnBase::SyndromeProvider$syndromes()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instanciate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderCountry$new(
  country,
  season,
  use.fever.level = T,
  unknown.has.fever = T,
  min.syndrome.count = 1,
  pain.age.limit = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{country}}{country code}

\item{\code{season}}{season number}

\item{\code{use.fever.level}}{logical use fever level variable}

\item{\code{unknown.has.fever}}{unknown fever has fever}

\item{\code{min.syndrome.count}}{count of symptom to take into account for some definition}

\item{\code{pain.age.limit}}{age under wich to exclude pain from definition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute"></a>}}
\if{latex}{\out{\hypertarget{method-compute}{}}}
\subsection{Method \code{compute()}}{
Compute definitions for all syndromes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderCountry$compute(weekly, intake)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weekly}}{weekly data}

\item{\code{intake}}{intake data with at least 'person_id', 'age' column}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderCountry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
