% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome_provider_covid.R
\name{SyndromeProviderCovid}
\alias{SyndromeProviderCovid}
\title{Syndrome Provider with COVID-19 like definitions}
\description{
Syndrome Provider with COVID-19 like definitions

Syndrome Provider with COVID-19 like definitions
}
\details{
To run need weekly with get_columns_for_incidence() column list
}
\examples{
\dontrun{
 season = 2017
 weekly = survey_load_results("weekly", get_columns_for_incidence(), season=season)
 intake = survey_load_results("weekly", c('timestamp','date.birth'))
 weekly = recode_weekly(weekly)
 intake = recode_intake(intake)
 provider = SyndromeProviderCovid$new()
 r = provider$compute(weekly, intake)
}

}
\section{Super class}{
\code{\link[ifnBase:SyndromeProvider]{ifnBase::SyndromeProvider}} -> \code{SyndromeProviderCovid}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pain.age.limit}}{age under which pain & headache will exluded from defintion}

\item{\code{use.sudden}}{logical, use sudden feature to build the syndrome}

\item{\code{definitions}}{character vector of definitions to output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SyndromeProviderCovid$new()}}
\item \href{#method-syndromes}{\code{SyndromeProviderCovid$syndromes()}}
\item \href{#method-check_definitions}{\code{SyndromeProviderCovid$check_definitions()}}
\item \href{#method-compute}{\code{SyndromeProviderCovid$compute()}}
\item \href{#method-clone}{\code{SyndromeProviderCovid$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ifnBase" data-topic="SyndromeProvider" data-id="compute_age">}\href{../../ifnBase/html/SyndromeProvider.html#method-compute_age}{\code{ifnBase::SyndromeProvider$compute_age()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ifnBase" data-topic="SyndromeProvider" data-id="is_sudden">}\href{../../ifnBase/html/SyndromeProvider.html#method-is_sudden}{\code{ifnBase::SyndromeProvider$is_sudden()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
instanciate object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderCovid$new(pain.age.limit = 5, use.sudden = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pain.age.limit}}{age to take into account of subjectives symptoms}

\item{\code{use.sudden}}{use sudden appearance of symptoms in the definitions, if FALSE will be considered as always sudden}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-syndromes"></a>}}
\if{latex}{\out{\hypertarget{method-syndromes}{}}}
\subsection{Method \code{syndromes()}}{
Get available syndromes names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderCovid$syndromes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character()
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_definitions"></a>}}
\if{latex}{\out{\hypertarget{method-check_definitions}{}}}
\subsection{Method \code{check_definitions()}}{
Update definitions list to compute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderCovid$check_definitions(definitions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{definitions}}{character vector of definition name to compute}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute"></a>}}
\if{latex}{\out{\hypertarget{method-compute}{}}}
\subsection{Method \code{compute()}}{
Compute definitions for all syndromes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderCovid$compute(weekly, intake, definitions = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weekly}}{weekly data}

\item{\code{intake}}{intake data with at least 'person_id', 'age' column}

\item{\code{definitions}}{character vector of definition to use}

\item{\code{use.sudden}}{logical if TRUE use is_sudden, otherwise consider it's always TRUE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame with each computed syndrome in column, and "id" column from weekly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderCovid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
