% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome_provider_2019.R
\name{SyndromeProviderRS2019}
\alias{SyndromeProviderRS2019}
\title{Syndrome Provider with 2019 revised ILI & Ari definitions}
\description{
Syndrome Provider with 2019 revised ILI & Ari definitions

Syndrome Provider with 2019 revised ILI & Ari definitions
}
\details{
To run need weekly with get_columns_for_incidence() column list
}
\examples{
\dontrun{
 season = 2017
 weekly = survey_load_results("weekly", get_columns_for_incidence(), season=season)
 intake = survey_load_results("weekly", c('timestamp','date.birth'))
 weekly = recode_weekly(weekly)
 intake = recode_intake(intake)
 provider = SyndromeProviderRS2019$new()
 r = provider$compute(weekly, intake)
}

}
\section{Super class}{
\code{\link[ifnBase:SyndromeProvider]{ifnBase::SyndromeProvider}} -> \code{SyndromeProviderRS2019}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pain.age.limit}}{age under which pain & headache will exluded from defintion}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SyndromeProviderRS2019$new()}}
\item \href{#method-compute}{\code{SyndromeProviderRS2019$compute()}}
\item \href{#method-clone}{\code{SyndromeProviderRS2019$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ifnBase" data-topic="SyndromeProvider" data-id="compute_age">}\href{../../ifnBase/html/SyndromeProvider.html#method-compute_age}{\code{ifnBase::SyndromeProvider$compute_age()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ifnBase" data-topic="SyndromeProvider" data-id="is_sudden">}\href{../../ifnBase/html/SyndromeProvider.html#method-is_sudden}{\code{ifnBase::SyndromeProvider$is_sudden()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
instanciate object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderRS2019$new(pain.age.limit = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pain.age.limit}}{age to take into account of pain & heacache}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute"></a>}}
\subsection{Method \code{compute()}}{
Compute definitions for all syndromes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderRS2019$compute(
  weekly,
  intake,
  definitions = NULL,
  use.sudden = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weekly}}{weekly data}

\item{\code{intake}}{intake data with at least 'person_id', 'age' column}

\item{\code{definitions}}{character vector of definition to use}

\item{\code{use.sudden}}{logical if TRUE use is_sudden, otherwise consider it's always TRUE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame with each computed syndrome in column, and "id" column from weekly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyndromeProviderRS2019$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
