% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndrome.r
\name{syndromes.prettify}
\alias{syndromes.prettify}
\title{Prettify syndrome names from InfluenzaNet DB health status name to another set
by default, transform to pretty names (for R) from the same set ()}
\usage{
syndromes.prettify(x, pretty.set = "influenzanet.2012")
}
\description{
Prettify syndrome names from InfluenzaNet DB health status name to another set
by default, transform to pretty names (for R) from the same set ()
}
