% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.r
\name{week_stamp}
\alias{week_stamp}
\title{Compute a "Week" stamp (like timestamp) : week number from 1970-01-01}
\usage{
week_stamp(yw)
}
\arguments{
\item{yw}{yearweek value}
}
\description{
This provide a continuous index for week number which not depend on period bound (unlike make_week_index)
This index is usefull to compute on weekly data (+1 is always the next week) without using date and to plotting
weekly data
It is usefull to plot week based data.
}
